[roll] Roll fuchsia [network-manager] Return futures to the eventloop
Let the eventloop explicitly join on the futures returned by workers
instead of spawning the workers' futures straight to the Executor. This
allows the eventloop to observe workers' errors (and potentially take
action in the future).

This change nows result in NetworkManger exiting with an error if any of
the workers terminate early, with the exception of the Overnet worker as
the Overnet service may not be available in all builds. Other than this,
there are no behaviour changes.

Test: Existing tests pass.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/392330
Original-Revision: f998e459ddb1442e54f125ae98196f054057d75f
GitOrigin-RevId: c5c66b3d11c41c0f262aadc326177a8224563388
Change-Id: I6974ab2911913c371f98c68b5c92b60b60c91bd7