[roll] Roll fuchsia [hwstress] Avoid status lines being overwritten on crash.
Currently, we write status lines as follows:

  "foo\r"

where "\r" will cause the cursor to move to the left-hand-side, allowing
us to replace it later.

Unfortunately, this means on a binary crash, the most recent status line
is lost, because the shell prompt (or any other output to stdout) will
overwrite it.

Update the status line to emit the "\r" when clearing a line instead,
meaning that the most recent status will still be shown on a binary
crash.

Original-Bug: 51547
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/392975
Original-Revision: 40ac9d965f412984dbf82ceb17eda52c7337b0d5
GitOrigin-RevId: f53fc4ef30fbe2c834e5f19f3d03bc38368066dd
Change-Id: I562b586414458057877e5f442ffda435da2d6cb1