[roll] Roll fuchsia [hwstress] Allow CPU workload functions to determine when to exit.
Currently, the |CpuStressor| class calls a workload function in a tight
loop. This makes it hard for the workload function to perform any setup
or teardown.

We adjust the workload function API to accept a "StopIndicator", and
make the workload function itself responsible for determining when to
stop.

This allows for workload functions to perform set up / clean up, such as:

  void DoWork(const StopIndicator& indicator) {
    // Set up
    void* foo = malloc(...);

    // Main loop.
    do {
      // (do work)
    } while (!indicator.ShouldStop());

    // Clean up
    free(foo);
  }

Original-Bug: 51547
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/392978
Original-Revision: 66becca2e5c586d9629f94ebb6e52a942813950d
GitOrigin-RevId: bb23b142163d759fc7bc72706070c89f75d3bd94
Change-Id: Ic6e95e2300b8d5a1979a6092878558bd759ea437