[roll] Roll fuchsia [fidlc] Handle diagnostic spans indicating EOF in JSON output
Currently, diagnostics with a span indicating the end of source file
cause an error when fidlc is invoked with --format=json. The JSON
diagnostic printer includes both the start position and end position of
a span, unlike the text printer. On a span that points at the end of
the file and that is of size 1 (it points to the null-terminating
character), this causes fidlc to read past the end of the string to get
the end position of the span.

This CL makes an exception in diagnostic printing for spans that begin
at the end of the source file.

Test: fidlc host tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/393271
Original-Revision: 0db58e66ac01e908a0eb7cd3c279d6e8f18a73bd
GitOrigin-RevId: 0a7689d1fe36616b1ec580f188275888b6a3d00f
Change-Id: I5a40c0eb466287f76cf69c513b8b9d90c6660f84