[roll] Roll fuchsia [bt][hci-emulator] Fix use-after-free in LE connect
Fix a use-after-free bug (detected by ASAN) in the bluetooth
hci-emulator where we posted a delayed task that closed over a pointer
to a peer that is not guaranteed to live long enough. Instead close
over a weak pointer to the controller and the peer address and then use
those to re-acquire the peer reference (if it exists).

Original-Bug: 49570
Test: bt-integration-tests

Not using a Multiply directive as there is currently at least one other
flake in bt-integration-tests that needs remedying, but this issue is
the more prominent one and needs an urgent fix.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/392233
Original-Revision: c7e7f6a86619faee8498b23114edce286c1456e0
GitOrigin-RevId: 42e267c5cbb20efce35968bf1271033624ad2d62
Change-Id: I3ab8f894c92838de6e3a7e97ad7c304e5adf6729