[roll] Roll fuchsia [crashlog][tests] Add basic integration test for kernel crashlog
Add an integration test that verifies basic behavior of stowing and
retrieving the kernel crashlog.  The test boots a qemu instance,
triggers a crash, which generates and stows a crashlog.  The system
then reboots and on the subsequent boot the crashlog is retrieved and
checked (only the most basic checks are peformed).

Updated the qemu integration test package to support arm64.

Changed "wait for qemu to start up" logic to wait for a newly enabled,
arch agnostic qemu trace event.  Previously, we waited for the
"SeaBIOS" message, which is x64 specific.

Currently, this test only works on arm64.  It's skipped on x64.  See
fxb/52843 for details.

A future CL will build on and extend this test infrastructure.

Original-Bug: 52182
Test: out/core.arm64/host_x64/kernel_crashlog_tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/391509
Original-Revision: 70e89419c226d036fef76ed98de3342c83c150fe
GitOrigin-RevId: 4534691be8c9079900d1e9996a9579d5ba372514
Change-Id: I405c15d57056e9a8f9f8b1f78c715c18f79f4e79