[roll] Roll fuchsia [camera][controller] Ensure that driver callbacks are posted as tasks on the loop
Prior to this CL, the callbacks from drivers like GDC/GE2D and ISP
executed partially on the drivers thread and partially on the
controller's loop.

This CL ensures that the callbacks are posted as tasks on the
controller's loop.

Test: Added asserts which fails without this change while running
config cycling on camera-gym for a long time.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/393156
Original-Revision: e0cac515fb76ed0b1b3f4b8047be5e682f2d20d5
GitOrigin-RevId: d4674ff38a262cdb1e1b863aacd696209140e81f
Change-Id: I0bba9adb99f42d15befadc43f3731f2f067ad956