[roll] Roll fuchsia [a11y] A11yManager waits to be fully-initialized before requestingg settings.
This change prepares the A11yManager to initialize the Screen Reader with its
correct locale. Because it is not possible to init the Screen Reader without
this information, A11YManager waits to receive this info before requesting
settings.

Requesting settings when the A11yManager is not initialized may lead in a race
condition between receiving the locale and receiving the settings.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/393500
Original-Revision: fb0d6b3bc5b19ae1b97a3d4b214e52193b5a497d
GitOrigin-RevId: 98d69bb4f03c4d54773bb55ef7050ab2f485d337
Change-Id: I2bb50622e3756bdf7dae7fdb3d9f462728add354