[roll] Roll fuchsia [omaha client] Add "extra fields" to the request apps
Add the ability to set extra fields on the app element of omaha
requests.  This is to enable non-standard features that are implemented
outside of the core library, such as always including the current
update channel that the device is using.

In the omaha-client service, set the current channel as one of these
extra fields, so that it's available for omaha rules processing when
changing channels.

Original-Bug: 52980
Test: fx test omaha-client-lib-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/393553
Original-Revision: 95a15ab3a9e3afe7c7d1815ac88f0f307a30bf9d
GitOrigin-RevId: d615365b8b585ac08229bd32ba185e94a4a9aa55
Change-Id: I319c111f20809e9f9d41eb1b7982c83a393ba68a