[roll] Roll topaz [dart] remove usage of snapshot_analysis.dart
This is a private file inside of the dart repo. The output files
are no longer used in Fuchsia and we are trying to move away from
mapping in the dart sources.

If this file needs to be used in the future it should be provided
as a prebuilt.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/topaz/+/394158
Original-Revision: c5fcc1a519a287cb1d8d2138a9e4b6b8292cbf4e
GitOrigin-RevId: 84bf43dac980e049e660c58914bb6978ac668b4e
Change-Id: I148cb71c9e8aba2fd51af580a8a51db43aa26687