[roll] Roll fuchsia [pager] Supply uncommitted pages
Extends the semantics of zx_pager_supply_pages to allow the aux_vmo to
not have to be fully committed. Pages that are not committed are
notionally zero anyway, and therefore have a well defined value in the
target pager_vmo. This can be done efficientally without allocating
pages in the target through the use of markers.

Allowing uncommitted pages in the aux_vmo provides flexibility in how
the kernel chooses to represent VMOs, and specifically allows it to
merge committed zero pages back to the global zero page. Without this
change there is no way to reliably provide a page of zeros by supply
pages, as you are racing with the kernel zero page scanner.

Original-Bug: 52652

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/392980
Original-Revision: 9528792aea3e361192afa6cd3767cd61befb8505
GitOrigin-RevId: d3d2b21abee49bad932415136432c407d0aca1dc
Change-Id: I83f92aef91bf4c69de55a04d352fe7ea12344b70