[roll] Roll fuchsia [kernel][timer] Make TimerQueue::TransitionOffCpu more symmetric
Originally, this transition code was a function which took two
cpu_num_t parameters to identify the source and the destination of the
transition.

This patch restores the parity between the target (now, the this
pointer on the member function) and the source (the parameter) by
identifying the later directly, rather than via a cpu_num_t.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/394459
Original-Revision: b1e0fe8c3b229760ab836174764ed4eac35ae7f5
GitOrigin-RevId: 1873b59a0cefd458c88a7e9d625b74b808a4b70a
Change-Id: Id8326538162a1ee5a7573a1bc5120cb25ff461d1