[roll] Roll fuchsia [ffx] slightly improve internal error output
All errors that are not handled by higher level subcommands, that
reach main will be considered a bug. We format the error output a
litlte more, and include the causation trail, while also telling the
user that this is a bug.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/393818
Original-Revision: 0d2a51b39a475003711004bfd72d733c5de7fce3
GitOrigin-RevId: 5e36951900776ff3870f4b4e3d3edbd18550c765
Change-Id: I9cf61631b21a9b87c1187130084139b0abda3b43