[roll] Roll fuchsia [cmc] Replace cml::Ref with contextual types
This change adds a procedural macro to generate reference types, since
there are a lot of them. It replaces the `Ref` type with these types. We
introduce an `AnyRef` type which is similar to `Ref`, but it is not
deserializable: it is only for use in helper functions that want to
operate on multiple reference types.

This change is a first step towards replacing cml_schema.json with serde
validation. Instead of validating the references with a regex, we can
rely on the serde deserializer to report an error. This change doesn't
enable that validation yet, though.

Original-Bug: 52963

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/392302
Original-Revision: 32966601484465d25d44ec327bbea5af87df7e9e
GitOrigin-RevId: 70f593887ece1b8e154a9d746325c3ea80d8a970
Change-Id: I2f132d92f4023fb3e3138d3907b25e7e1a3ed479