[roll] Roll fuchsia [oom][logs] Drop "OOM" prefix from debug logs
All the logs we print from the oom-thread have the "OOM" prefix, which
can be misleading. The presence of the word "OOM" seems to indicate that
the we're running out of memory, which can be confusing if the log line
printed out was "OOM: memory availability state Normal".

Replace the "OOM" prefix with a more neutral "memory-pressure". Also
rename the "oom-thread" to "memory-pressure-thread".

Test: oom_tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/394793
Original-Revision: 5e3ec272cba73310062396f0eca490c9e274ef09
GitOrigin-RevId: 91e114d2ab42b420d97b1193b806607113a8aaaf
Change-Id: Ie74f2a7263b1eb59abfa0479b03707dc7d47e471