[roll] Roll fuchsia [fidl][llcpp] remove Initialize indirection
This initialization is not needed because we just allocated the message.

This doesn't change the codesize because the compiler was smart enough
to see two initializations and to just keep one.

This cl will allow to remove some templates later.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/394739
Original-Revision: 9b4b9d8fedb16bc99c045793ecd263e2d7a60740
GitOrigin-RevId: 02c015879489ee03106dd96d25971cbdfea5d945
Change-Id: I93eface024a29d170386f07c2eeb18a9106edddd