[roll] Roll fuchsia [brightness_manager] Fix bug with blue/green swap
The blue and green values were swapped on the service side
of the fidl interface to conform to HID spec, but broke the
parsing on this side. The blue green values are swapped now
in the parsing logic to correct the mismatch.

Test: fx test -o bright_manager_tests
Original-Bug: b/157594412
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/394693
Original-Revision: 20edc6d4dc51383b38483a6b829d134f8262b8db
GitOrigin-RevId: 2b9812ebfde91526dbb744c021cdb0ce42b9af3b
Change-Id: I089886fbe68e6cda100d7df4057dab9bd9be7c04