[roll] Roll fuchsia [flatland] Convert matrices to rectangles.
The conversion logic from matrices to rectangles was moved out of the
Renderer in change 391533. This change formalizes some of that logic
into a function to be called in the main render loop.

The intermdiate matrix format is not currently needed anywhere, but
will likely be necessary in the future for transforming input, so
this change leaves compute matrices and computing rectangles as
separate functions.

Rectangles are still computed and used for LinkSystem::UpdateLinks(),
which ensures that the Flatland API calls result in matrices that are
valid for rectangle conversion.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/393067
Original-Revision: df8ce38640b9684c4555382beb127486715a8ad7
GitOrigin-RevId: 36d8bb25654890dc7f5708fca188018d33e63738
Change-Id: Iab78fe5e06569dfd29952b3c8ca2233375073bec