[roll] Roll fuchsia [fidl][llcpp] RawMessage intermediate class
Defines a new class RawMessage which holds a BytePart and a HandlePart.

Uses it inside EncodedMessage<FidlType>.

This way, for some functions, we will be able to use a RawMessage
instead of an EncodedMessage<FidlType>. That means that we will be able
to transform some templates into functions (which means being able to
factorize the code).

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/395020
Original-Revision: 1803b47d8a0e8866f805d1a7a2cdc2dfe4b15919
GitOrigin-RevId: 08d2bb7cfb8e7aadde3f285ce12441f33825dae6
Change-Id: Ie7729e9f2315acf5c64104b990929214b1fd3d8c