[roll] Roll fuchsia [packet] Clean up documentation
Documentation has gotten somewhat out of date over the course of a
number of changes. This commit fixes a number of documentation
mistakes, and adds some clarifications.

Notably, this change removes the "Monomorphization Overflow" section
from the crate documentation, as the methods that it documents have
been removed. We can add it back in the future if those methods are
added back.

Test: No behavior change

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/395023
Original-Revision: 5dc34d8da452527d299a164796e3ea235ddbb097
GitOrigin-RevId: 2809f663ce52d70e1f4ffa7b49de5d23763912a1
Change-Id: I21d0539d0941340b8a6f82586ff3645b00ff12e4