[roll] Roll fuchsia [fvm] Generate fvm image for flash use.
Add support to generate an fvm image that is suitable for flashing
in fastboot. It's currently tested for sherlock.

A few notes:

Fuchsia currently generates a raw fvm image "fvm.blk" for qemu use.
The size is only about 480MB, which is actually flashable without
android sparse image support. However, the disk size specified in the
metadata header of the image is the same as its current image file size
and not the fvm partition size of the target board. This has been
observed to cause ZX_ERR_NO_SPACE errors during netsvc paving when the
fvm partition is pre-written the raw image. As discussed, it is
necessary to either generate the image with the correct disk size in the
first place, or extend it to the corrct size with fvm tool later. The
fvm tool does two things in these operation.

1. Write/adjust metada header, slices according to the specified disk
size;
2. Resize the image file to that of the specified disk size.

For sherlock, the fvm partition is around 3.3GB. The resulting image is
too large to flash to the device without android sparse image support.
But it seems that it is possible to only perform 1. but not 2., which
still allows the image to specify the correct disk size but without
increasing the image file size. This is the solution the CL is currently
using. The image is the same size, around 480MB, and flashable without
android sparse image support.

The generated fvm image does not come with the ssh authorized_keys by
default, without which most of the fx host side will fail.

The CL currently generates dedicated flash images instead of modifying
and reusing the existing fvm.blk, data.blk etc. This is for avoiding
unexpected impact caused by changing those images already in use
somewhere.

On sherlock, the generated image can be flashed by

"fastboot flash fuchsia-fvm <image>"

It takes about 100 seconds.

Original-Bug: 48803

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/389653
Original-Revision: 18b2fb99cb199674af2a7223605bd38062c28e5e
GitOrigin-RevId: 635e65faf3b587c4df8e09c2e2d6023e4057fe7d
Change-Id: I0b385d681bcda2d2e7080710f5800051c1fc3b29