[roll] Roll fuchsia [pci] Extend BusLinkInterface into BusDeviceInterface
Devices need a way to request the bus driver allocate
their MSI objects, so extend BusLinkInterface to be
BusDeviceInterface and now reflect an interface for
devices to request resources and device linkage from
the parent bus. This is also required for GetBti which
is addressed in a later CL. Pulling BDI out of the bus
header was necessary to allow a proper dependency chain
with the fake test objects.

bug: 32978
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/393734
Original-Revision: 83518f58315ad2d073a9ff0b6549e633fa444c19
GitOrigin-RevId: cfc81662a5386b885f73b58a5572a308717e8c41
Change-Id: I53d53d289791415376bc5eb2e6190aa8624a4328