[roll] Roll fuchsia [carnelian] Refactor Scenic support
Refactor scenic support as a general cleanup and to more easily add
better frame scheduling later.

The goal was to move all scenic specific code into scenic modules. This
CL doesn't quite get there, do to the need to forward scenic input
events to the app strategy for processing, but it is a long way towards
it. Refactoring the input handling feels like too big a step for this
CL.

Moved the app strategies into their own modules to make the separation
easier to maintain.

Got rid of the supports_scenic method on AppStrategy as it is no longer
needed. Instead there is a method for creating a view at launch time
for testing. This is only implemented on Scenic, as Framebuffer always
creates a view at launch.

Initialization order has changed enough that I was able to remove the
unnecessary use of Option for any fields in App. That structure is now
fully initialized on creation.

ViewController now has a single constructor function which uses the
view strategy for behavior differences on Scenic and Framebuffer.

Got rid of the ScenicResources struct, as it no longer needs to be used
from multiple places. Now its fields are part of ScenicViewStrategy.

Removed the unnecessary _async name suffix on functions that used to
have a sync and async version.

Testing: all automated tests, plus manual testing of all examples,
recovery and terminal.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/392808
Original-Revision: 86ff701cac310075a78e85a5bfeec975e4225e48
GitOrigin-RevId: 2c5096d43bb916c1eb1c4485e9afdb9478aaf984
Change-Id: Ib21c2fdd94001d2fe17a373afd594a4ce53de8ee