[roll] Roll fuchsia [sme] [brcmfmac] Handle channel change indication for SoftAP.
Channel change indication is issued when the firmware changes
the SoftAP's channel (for example, the client interface in the same
device associates to an AP running on a different channel).

Test: Added a unit test to check the handling of this event.
Added another test to verify softap creation failure in the driver,
but this test needs the sim timer changes.
Also, verified on astro and sherlock that the channel indication is
handled. The test sequence is:
- Start SoftAP on channel x
- Check wlan status reports SoftAP is running on channel x
- Associate client IF to an AP on channel y
- Check wlan status reports SoftAP is running on channel y

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/383932
Original-Revision: 36881ea77dbd39f3e11f6cdc9a673a98b0f441fb
GitOrigin-RevId: 364af1446eada39f1279750c005c15d22ec1b687
Change-Id: I965c16d6b8d392240695a87f4dd80e63f0793916