[roll] Roll fuchsia [covargs] Remap source files to new checkout.
The coverage recipe creates a new checkout to run with the covargs tool
which has a different path than the checkout that was actually used to
build. We need to remap the file paths to use the new checkout so it
can properly find the source files and upload their corresponding html
files with the coverage report.

Test: https://chromium-swarm.appspot.com/task?id=4c95a9df8a55d410

Original-Bug: 49473
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/394179
Original-Revision: 5267492ea3e595ffbc0869b4bd6275c22a6f2565
GitOrigin-RevId: b5bef661a6d47f440628d0bd3b87153d642c126a
Change-Id: Ide652ab5b108e92b9249358e2d60d4f162652d8e