[roll] Roll fuchsia [bt][a2dp-source] Select correct bitpool for SBC
SBC negotiation would select 53 as a constant bitpoolvalue whenever we
selected SBC, which happened occasionally.  It is not required that
sinks support this value, and sometimes have a smaller maximum value.

Select the maximum value that is supported if it is below 53.  Also
change the default value to 51, which is recommended for Joint Stereo
48kHz audio by the A2DP spec (Sec 4.3.2.6).

Original-Fixed: 52376

Test: updated test_stream_selection in bt-a2dp-source-tests

Multiply: bt-a2dp-source-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/393729
Original-Revision: eddca88043cd90ac5f5bcc42f170357891b687c6
GitOrigin-RevId: 4b75143ca869e642af37db3eb09fa4fede3f5205
Change-Id: If41fa344ee99c3c572cceed902dd5ba91f49c73a