[roll] Roll fuchsia [ffx] Begin internal documentation.
This is a good start.  There still needs to be some documentation around
the persistent configuration files, how to setup an environment, etc.
Also, how to use the config macros in the library.  But since this was
timeboxed, this is what I have so far.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/394187
Original-Revision: 5c2feaf302dcbad339c13cf18b3da58e8c6ff003
GitOrigin-RevId: 3954f02a8b19ec5ce97941eb35ddc0d4e4bf6911
Change-Id: I507ba3e70839d04a715ba6c7f0814677f8de5d07