[roll] Roll fuchsia [ffx] Detect DaemonProxy param in ffx_plugin
With the double-dispatch it's now possible to identify that a ffx_plugin
takes a DaemonProxy instead of a RemoteControlProxy.

Extending this method in future CLs should allow for ffx_plugins that
take a number of arbitrary Proxies, and then use the new RCS selector
interface to gather the proxies - generating all the boilerplate code.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/393195
Original-Revision: 226db2dbc5b4d4020b9dd6704043796d276dfec5
GitOrigin-RevId: 9b7767714417fb4eecd89b2ea0238969a59108e3
Change-Id: I50ff079779cc809d98f323b0013f746b20106c93