[roll] Roll fuchsia [settings] Add InspectBroker to write all setting values to inspect
The InspectBroker sits between the registry and setting handlers. It listens for the Restore command on boot and asks each setting handler for its value when it replies to the Restore command. When a setting handler tells the registry its value changed, the InspectBroker also requests the value. This isolates all the inspect-related logic in a single broker and no other pieces of the settings service need to worry about it.

Original-Bug: 51123
Fixes: 44874
Multiply: setui_service_tests
Test: fx test -o setui_service_tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/384690
Original-Revision: 63295a33826f2c91ec3d45cb6f81bcded595beb6
GitOrigin-RevId: 761be550d84f6520864e0b1f54034f424c7fbcab
Change-Id: I841014429fd5b4fa47f7db8f0dd6e43ed5045fdd