[roll] Roll fuchsia [devices] Do not explicitly install libdriver in bootfs.
Now that no bootfs content comes from the ZN build anymore, libdriver always
naturally gets pulled into zbi's via the inclusion of a driver which depends
on it.

Adding libdriver directly is problematic as doing so completely bypasses
variant selection which is done at the executable/driver level. The current
zbi building step does not rely on the dependency tree in order to identify
what should be included in the zbi, but its replacement will. The dependency
removed in this change would have had the consequence of always including
the no-variant version of libdriver (and some shared libraries it depends on)
in zbi's, which is inaccurate and wasteful.

Original-Bug: 45680
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/395773
Original-Revision: 1d207c5243a982d6baea93054cb90c1cfc5a2879
GitOrigin-RevId: 4bdb858fd5f0b6256f29cc969b63c138edab4b1e
Change-Id: I1516a4f369c5dfa3c094004f475af1629d754c34