[roll] Roll fuchsia [fbl][lists] Add split_after to lists.
Add a split_after operation to the single and doubly linked list
containers.  The operation accepts either an iterator or direct
reference to an element in the list (call it A), and splits the list
in two lists at the point after A.  It will return a new list whose
head is the element immediately following A, while A will become the
tail of the list which was split.

The split_after operation is O(1).  Sized lists do not support split
after as maintenance of the size member would require O(n)
complexity.

Tests: new unit tests
Original-Fixed: 53051
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/396033
Original-Revision: 17614828b0b9423df5f4ece2f07625fd0e7c9885
GitOrigin-RevId: c77d911b1ac6182775793d316579be3c9a9004c4
Change-Id: I5ad5887dce7f23ca8aee7ec999c963d4ab3d6ef0