[roll] Roll fuchsia [blobfs][pager] Pass const reference to UserPagerInfo
The user pager functions need to consult a blob's |UserPagerInfo| for
various bits of information required on the paging path. Instead of
passing in a |UserPagerInfo*|, pass in a |const UserPagerInfo&|. Once
the |UserPagerInfo| is created at the time of initialization of the
blob, it is never modified. Besides, these functions should not be
tampering with the contents of the |UserPagerInfo| anyway.

Test: No behavior change. Existing blobfs unit & integration tests pass.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/396055
Original-Revision: f51f1fe280c53982598b72e7d03dd04838cd9264
GitOrigin-RevId: fcae249b7682c290210995a7da4e185ad69def3a
Change-Id: I20794f48b6435db2e720f456099ec8cdf797e1fb