[roll] Roll cobalt Removes contents from the file observation.proto.
In Cobalt 1.0 we switched from observation.proto to
observation2.proto. Most of observation.proto was
only relevant to Cobalt 0.1, but we needed two
messages from there: BasicRapporObservation and
ValuePart.

In this CL we move those two messages into
observation2.proto and delete most of the contents
of observation.proto in preparation deleting
the file.

In a follow-up CL we will delete ValuePart. This
is part of an older interface to the Basic RAPPOR
Encoder that is not relevant to Cobalt 1.0.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/cobalt/+/382740
Original-Revision: 029a08443d012058ca4ce55f2658bdef800d7f9d
GitOrigin-RevId: 47834e2be985c1388da0f9d5e1a9367c11736362
Change-Id: I917948c61ebb1bc615c191b80ac7f91a807cc7c8