[roll] Roll fuchsia [bt][host] Fix inquiry scan interval
On both Intel and Broadcom controllers the bt-host driver always logs the
following warning:

[gap-bredr:bredr_discovery_manager.cc:503] write inquiry scan activity failed: [status: invalid HCI command parameters (HCI 0x12)]

This happens because the controller rejects the parameters given to the Write Inquiry Scan Activity
command during bt-host initialization. This is due to a mathematical error, in which the intended
300ms interval maps to 0x1E0 in jiffies but was incorrectly coded as 0x1E1 which does not map to a
whole number in milliseconds.

Test: 1. bt-host-unittests.
      2. bt-host should no longer log the message mentioned above.
      3. Discoverable mode should continue to work as expected.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/396070
Original-Revision: 4b3ac75934dd8904955b3e2b8f651ea7acf0c5a6
GitOrigin-RevId: 09b3d4314919be3a80e0c92ed3533de1b733ef89
Change-Id: I154338eb0bfd884be8ccaf42d58a3520f609f339