[roll] Roll fuchsia [a11y] Fix for memory leaks in annotation view tests.
This change should fix the memory leaks in the annotation view tests. I
believe the issue was that several of the derived classes in the
annotation test file did not override their parent classes' virtual
destructors. My understanding is that only the parent class destructor
is called upon destruction of a derived class object in this case,
so the derived class' resources are never freed.

BUG: 46525
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/369656
Original-Revision: 2812e69afc1a0d5a303df085975356b260a1baae
GitOrigin-RevId: 3f2bc92b553450e22a6e5666215502364eab1449
Change-Id: I5c05f8b1b082e86ea89d70da2a8562b2b7a24ee8