[roll] Roll fuchsia [soundplayer] tolerate wav files with extra bytes
Other players are apparently OK with a few extra bytes attached to
the end of a WAV file. This CL changes the WAV parser so it will
tolerate up to 3 extra bytes at the end of the file. If there are
four bytes, they're expected to be a RIFF tag.

Test: fx test soundplayer_bin_test
      Three new unit tests added for this case.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/396213
Original-Revision: 33a538578421f0c4464fb0aa089e3c797a2ad0e2
GitOrigin-RevId: 4d6ac7d7ed437706906fd8b39ce8dc80ce22b900
Change-Id: Ia8aa95cac7941a8aa1c64a6ae6d9a73f4781067f