[roll] Roll fuchsia [kernel][pmt] Fix a pinned memory leak in the kernel.
Fix an error in the kernel where it would follow an error path out of
the pin syscall _after_ having created a valid pinned memory token.
When the PMT kernel handle went out of scope, the dispatcher (which
never got handed back to the user) hits on-zero-handles, but has
memory still pinned, so it gets added to it's BTIs quarantine list
effectively leaking the pages.

re: Testing.  This _should_ have been caught by core-tests, but was
not.  The VmoTestCase.CompressedContiguous has a test where it attempt
to pin memory, but does not provide enough space to store the
resulting scatter gather list, and expects to see ZX_ERR_INVALID_ARGS
as a result.  In this case, it does, so it appears as if the test has
passed.  When the test ends, however, the BTI goes out of scope and
ends up printing the (now non-empty) quarantine list in the klog.

Moving forward, it would be best if we could modify the host side
core-tests zbi infrastructure to look for strings like this one and
consider it to be a test failure if it sees one.  Alternatively, we
could modify the actual warning printed by the kernel to look more
like an ASAN error.  AIUI - these are already being scanned for in the
zbi test environment.

Testing: See above.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/395565
Original-Revision: 877f13b39c7b4506508d9455f92717d6da415fbd
GitOrigin-RevId: 3f6c5854b31b8f823234e4bd9ae2ccfeea693244
Change-Id: I68f706a1dcef24dfcf67a7693aeb6c997e244fa1