[roll] Roll fuchsia [lib][fake-object] Use zx::status and assert less
This refactor moves most of the internals of the library
to use zx::status and removes cases where asserts
would be used to handle cases where the method signature
made it difficult to convey success / failure. This brings
all of the FakeHandleTable methods in line from an api
perspective and reduces the amount of work needed to use
the library with derived types. More exhaustive testing
was added in the process.

bug: 32978
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/395370
Original-Revision: cb1275f45ddf8649e64c7efee198498ccd04e2c2
GitOrigin-RevId: c2b1543d10f3f2d2897d6deaf10e46627c3e2b3a
Change-Id: I42e231ee2845b47d2d4485fca600bc48fcce9dc2