[roll] Roll fuchsia [mediaplayer] fix mediaplayer flakes
This CL fixes some common mediaplayer flakes. We were seeing audio
packets being dropped, because the fake audio renderer wasn't getting
scheduled and failed to retire packets quickly enough. The mediplayer
code driving the audio renderer was demanding new packets based on
renderer lead time requirements and real time. Because it assumed
that the renderer would retire packets promptly, the payload VMO for
the renderer would become exhausted, causing allocations to fail and
packets to be dropped.

The new behavior refrains from requesting packets when a new packet
is likely to overflow the payload VMO. This will generally not occur
on real hardware, because the audio renderer doesn't get that far
behind. If that behavior does kick in, lead time commitments may not
be met, but the renderer is behaving badly anyway, and the user is
probably hear glitches already. Time is never slipped, so the
condition is likely recoverable assuming the renderer gets its act
together.

This CL also fixes an error reporting issue in the fake audio
and reduces the size of the payload VMO to 1s of audio data. The
size was bumped up recently in a vain attempt to work around
the overflow issue.

Test: fx test mediaplayer_tests
      Modified integration tests test the new behavior and verify
      there are no regressions.

Original-Fixed: 52622

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/394935
Original-Revision: f66146993e56c1616cf259d65cb5cf24e7d48736
GitOrigin-RevId: 4fb042fb3a7de69910c1cc4797705cd4397ee7b5
Change-Id: I54bf76ae8c73082b85a210fd58846542f75696f4