[roll] Roll fuchsia [kernel] Change format of internal log buffer
This adds a few features the archivist team has wanted: information
about the amount of data rolled out of the log when we don't read fast
enough, and log severity in line with the severity numbers currently
used in userspace.

Severity is now a separate item from log flags through most of the code
base, but we stop short of changing the syscall arguments for now.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/389975
Original-Revision: f4cd75a3896aac599159eef722cc3318e94a56c4
GitOrigin-RevId: bc68c3986cdf46968675d84187decebb98d61163
Change-Id: I09000594c1c1b3f54cedb0fbc0e696fc76319a94