[roll] Roll fuchsia [netstack] Avoid double closing TCP sockets
Move resource (and refcount) cleanup to a deferred function in FIDL
handler. This unifies the behaviour of fuchsia.io/Node.Close with the
behaviour of a closing zircon channel. Prior to this change, a FIDL
Close call would result in two calls to close, decrementing the refcount
twice.

Refactor the associated tests to more convincingly simulate the scenario
and more strictly check state over the course of the test.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/396477
Original-Revision: 7a7da2b4c058cf589eeae44c6078496da208348a
GitOrigin-RevId: a41ed7d9e063bcf6ea4b0ad09ba76a2705a29173
Change-Id: Ia19f1a6c838d4a576731cf15b3aef3be2878d04e