[roll] Roll fuchsia [fshost] Bump pkgfs startup deadline
We are investigating a possible performance regression in pkgfs in
which it takes ~30 seconds to start on emulators when running an SDK
companion image. Bump the timeout to 40 seconds while we investigate
both the performance of pkgfs startup and removing this timeout
entirely.

Original-Bug: 53639
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/396849
Original-Revision: e374ba5c407d435b6f371febd9e1ac46fd7718f1
GitOrigin-RevId: 5ebf9c93876ea6a1482797bb706d680b8f058ad0
Change-Id: If6f2c13d0721e4d0a428528aa8947c6008e71ba5