[roll] Roll fuchsia [build][fuzzing] Fix fdio build for fuzzer variants
libfdio.so was recently migrated from ZN to GN as part of
fxrev.dev/392329. Unfortunately, that CL broke the workaround for
fxbug.dev/38226. This wasn't caught by CQ as fuzzer builds have been
previously included in CQ. (See fxbug.dev/53411 for the request to
change that.)

This CL includes the workaround for fxb/38226 into
//build/images/finalize_manifest.py, so that the generated package
manifests correctly indicate to use the uninstrumented libfdio.so.

It also updates the "--fuzz-with" option to `fx set` to enable the
variant toolchain needed to build the sanitized-but-not-covered
libfdio.so.

Original-Bug: 38226, 53028, 53411
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/394198
Original-Revision: 596f6ede55947e08eae2303b00d41106cc2edafc
GitOrigin-RevId: c3bce9596fa05523e42ad6b29e3850e6a5a9ed45
Change-Id: I58f55277944c802b0bb0b2c4ad1829741d6e691c