[roll] Roll fuchsia [appmgr][cmx] Propagate all entries in "program" to runner
This allows component manifests to specify additional attributes in the
"program" field for the runner to use in executing the component. This
is closer to the component framework V2 definition, which passes all
fields in the corresponding dictionary to the runner to interpret. The
attributes are distinct from the "args" value, which are for the use
of the component being run.  Only string values are permitted to
reflect fuchsia.sys.ProgramMetadata definition.

This updates appmgr to propagate these fields through and updates the
cmx_schema to allow additional properties in the "program" object so
long as the values are all strings.

Original-Bug: 53243

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/396269
Original-Revision: 98bf447fbc66ef1b3e4d450f2c4df212eb8d8b1b
GitOrigin-RevId: bc016cd6c780d1928a964cbb9153af82e85996bd
Change-Id: I28cfd812d07ce91f5c6985b1bcbedc98957f623e