[roll] Roll fuchsia [audio][astro] Change PDM paramters for 96kHz
We change the PDM CIC filtering parameters so instead of halfing the
last downsample rate in the chain, we half the first one and adjust the
gain accordingly as seen in a reference implementation. This improves
the performance seen in capturing 28kHz tones in Astro, and shows no
adverse effect on lower frequencies.

We also update a HPF parameter difference from the reference
implementation.

This change is now virtually identical to Eric's
https://fuchsia-review.googlesource.com/c/fuchsia/+/386033

Original-Bug: 53207
Test: Manual recording with audio-driver-ctl with lower (16kHz tone and regular music) and higher (28kHz tone) frequencies.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/395676
Original-Revision: bdcedb7134bba809e63bc648b205a2036a54afe5
GitOrigin-RevId: 1879fd533e941eb6686d38b278d3883ad19683a8
Change-Id: Ia8ecf0ea57c7d991aedcb2aa986616591191dcc5