[roll] Roll fuchsia [e2e][fidlcat] Force permissions to 0600 for ssh key
Swarming stopped maintaining file permission bits, so when the private
key is isolate from the builder to the tester, it becomes 0666 which ssh
will reject. In order to work around this, chmod the private key before
running the test until we switch to not isolating the key.

Re-enables test that was disabled in
https://fuchsia-review.googlesource.com/c/fuchsia/+/395780.

Original-Bug: 53492
Original-Bug: 53613
Original-Bug: chromium:1092020
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/396842
Original-Revision: 9a5dd3269833991dbc7a22e84fc096f4ad962340
GitOrigin-RevId: 07848b28da284ab13b62891cfa0ba354ef4c3586
Change-Id: I109868f608d00779dad413a13f93a05f31c29611