[roll] Roll fuchsia [debugger] DieRef->SymbolRef, make Kind an enum.
Renames IndexNode::DieRef to IndexNode::SymbolRef, and changes the "is
declaration" flag in that structure to an enum.

This change was started to allow adding ELF symbols to the index. The
more pressing issue requiring that was solved a different way so I will
not be following up with more patches along those lines soon. But we
will likely want to add ELF symbols to the index in the future which
makes this generalization still desirable.

There should be no behavior change.

Original-Bug: 53091

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/392533
Original-Revision: bb791386c28e0056eb62f1d44915a1c18b9e28e4
GitOrigin-RevId: 567b9306851e6bce5be4b40150aa1b0ba3547d74
Change-Id: I802563246f777564911d5116dccd79295d2ed034