[roll] Roll fuchsia [scenic] Fix ViewTree error message for root view_tree
There's an error message in ViewTree that fires when a Session has more
than one connected view. Because the ViewTree also counts the Scene node
as equivalent to a view this message also fires if the root session
contains a view. This CL fixes that, allowing the root session to own
a view. Necessary for input injection.

Test: No functional changes. Manually tested.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/397034
Original-Revision: 32a92229f38ba502d8d21379d86efc38f4220f55
GitOrigin-RevId: e706364ffb70a1b14b5346ab14be8ea2548b9b2d
Change-Id: I38211cc815b89b0a3474713b33bc3f82372a473c