[roll] Roll fuchsia [zx] Fix zx::vmar argument order to match C interface
When we updated the C interface for zx::vmar to match the API rubric, we
did not update the C++ interface. This CL begins the process of updating
the C++ interface to match the C interface by introducing new methods
with the proper argument order.

For zx::vmar::map, we can add the new method with the final name because
the zx::vmo argument will disamiguate the overloads. For
zx::vmar::protect and zx::vmar::allocate, an actual parameter of 0 would
be ambiguous, so we need to go through a more complicated sequence of
introducing a protect2 and allocate2 with the proper argument order,
removing the old methods, and then migrating back to the original names
with the proper argument order.

Original-Bug: 52432
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/394178
Original-Revision: e8a5e89f1ea703d98d249b7b5592e26404c4b518
GitOrigin-RevId: 73df09c48e32492dec7d9a880e79a1ae04797ce9
Change-Id: Idacfbb8181720a496efcbfb8212cd2df6be1cbcb