[roll] Roll fuchsia [wlan][factory] Combine and increase buffer size
The factory tool uses the same buffer as input and output so it would be
eaiser if the size of request and result vectors match.

Also increase the size to 16k because it is the largest command buffer
defined anywhere in the factory tool.

Test: None. There is no functional change.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/396077
Original-Revision: b86bf4af6b685464497df6110d4a1a3c0b33a748
GitOrigin-RevId: c161eac8add5e9878f8b167261fe8d7cef3b9bd8
Change-Id: Ie51d3f0d0609856f840f8088a42ec3e349e95da7