[roll] Roll fuchsia [Fuchsia DevTools] add --once option to Harvester
The Harvester is normally a long running process. For some testing purposes it
is helpful to run the update loop one time and exit. This CL adds a command line
option for `--once` that will run the main update loop one time and then exit.

Bonus: adds a join on the secondary thread so that it shuts down without
segfaulting when the main thread exits. This bug was exposed by adding the
--once option.

Test: src/tests/end_to_end/harvester_test.dart
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/391756
Original-Revision: a1d498ecdbb94b3e5c5fa66af375934655f207fa
GitOrigin-RevId: 586f5fe6eb8e5ba05290055cadffc0192c7f6ff3
Change-Id: I8cd1029db5d66d1d42b34fe893a2d4c5d62a7b17