[roll] Roll topaz [dart] link platform files for flutter_test
We need to include the platform files for host tests since
a recent change in dart fixed a bug which was not actually
honoring this flag. On the host we do not manually link the
platform files like we do in the runners.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/topaz/+/396700
Original-Revision: 4ed89a113a07f244970672d1da072dc1b98734e3
GitOrigin-RevId: fc6ec6230df20b9b0a33ded542c75331ce8c9a95
Change-Id: I1ff3dea150dc147fa2fcb88d70631beb00c6d634