[roll] Roll fuchsia [audio][sherlock] Allow and warn ring buffer position too big
Assertion for ring buffer position too big from the HW added
in https://fuchsia-review.googlesource.com/c/fuchsia/+/378538
shows that it is possible to get a report from the HW that is
beyond the ring buffer size. This CL changes the assertion to
a warning and a reset of the position to 0.

Original-Fixed: 52004
Test: audio_driver_tests
Multiply: audio_driver_tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/392774
Original-Revision: 5018ddac75995f102530f9bc2ceabad41939c593
GitOrigin-RevId: 5dec0ba8f9a3c70db6e4abf448ccdd3bec99a524
Change-Id: Ic1d120840d8ec49718e767f65f528075c3a79a5c