[roll] Roll fuchsia [kernel][vm] Handle zero length VMO_OP_ZERO
Weakens the assertion in ZeroPartialPage to permit the start and end
offset to be equal. This can happen when the original length passed to
VMO_OP_ZERO was zero. ZeroPartialPage can correctly handle zero lengths
so this weakens the assert instead of adding a special case for zero
lengths.

Test: Included test triggers kernel panic without these changes

Original-Bug: 53484

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/396618
Original-Revision: 0170dfd65b0fa7acd17ee657daf2417d4b0707be
GitOrigin-RevId: 462180904c05863217efc0ea30edaeff255f0135
Change-Id: If626b0ca3293b62e83288f81dbece542d235233e