[roll] Roll fuchsia [affectedtests] Reland affectedtests speedup
Now with 100% less flakes.
This was previously flaky because the logic to shard the
Ninja graph in its dotfile representation wasn't stable.
As a result we would fail to acknowledge some of the nodes
or edges, and then miss some possible paths in the graph.

Multiply: affectedtests_tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/397044
Original-Revision: dfd71582c8d92c0929622d05b59cce45fb3aa630
GitOrigin-RevId: fe7da7b5573a4a7f65447852aafd8481ea38f3d4
Change-Id: I9f6da3f53311ae9adcaec8113c0108d8569ee232