[roll] Roll fuchsia [ulib][zxc] Add nodiscard attribute to zx::status and fitx::result.
The C++17 "[[nodiscard]]" attribute gives a warning if a function returns
a zx::status or fitx::result, but the caller ignores the result.
Ignoring the result is often indicative of a bug, where the caller
hasn't carried out error handling.

In cases where the error should be ignored, casting the result to void
will suppress the warning, as follows:

    zx::status<> MountFilesystem();

    void Init() {
      MountFilesystem();  // warning emitted

      zx::status<> status = MountFilesystem();  // no warning
      if (status.is_error()) {
         // ...
      }

      (void)MountFilesystem();  // no warning
    }

When the library is being compiled with C++14, we fall back to
__attribute__((__warn_unused_result__)) if possible, or otherwise just
ignore the attribute.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/395763
Original-Revision: b29f79a52f996d6c9aff0e3bdff247ed0b6a91cb
GitOrigin-RevId: dd506d3d7c3dbfee7821a1e4e78debe01ea37672
Change-Id: Ia9a755eeaee714e2e956272e5e09b0cf89b305fe