[roll] Roll fuchsia [cm][test_utils] Adds BlackBoxTestBuilder
With the need to add an additional argument to the launched
component manager for BlackBoxTest, some callsites would
need to be updated to pass an empty vector.

A better pattern is to use a Builder to configure the BlackBoxTest
without needing to specify all parameters.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/396107
Original-Revision: 0de6d252e4ed67a4f0fe363f841f9ba9549526c0
GitOrigin-RevId: f464b45da5367bb69f583022258fb4627ba29db2
Change-Id: Ie0ef2744aec2043105715c6a2e295d27722c975d