[roll] Roll fuchsia [mdns] remove check for config dir existence
A check was recently added to make sure the config directory exists
before calling JSONParser::ParseFromDirectory. As it turns out, that
method was supposed to handle non-existent directories without
indicating an error, so the correct fix was to fix the method. That
change is made by fxr/395654. This CL reverses the mdns change,
leaving the relevant test in place.

Test: fx test mdns_tests
      Unit test for the non-existent directory case remains.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/395673
Original-Revision: 234bf6b1af4095090e0527c7c56c41ed1df7d063
GitOrigin-RevId: 28dcb0222e81d454c564cf5affc24cdc531dc8c2
Change-Id: Ia76328eec1d9e104482b496bd2f0d89e8c5e3f96