[roll] Roll fuchsia [settings] Fix erroneous controller teardown logs
The logic was wrong for when to report an error during teardown. When
the teardown stage is reached, the entry for that setting type should
already be removed from the mapping. The teardown phase was logging an
error if it didn't see it in the mapping, expecting it to still exist.

Original-Fixed: 53458
Tests: fx -o setui_service_tests
Only effects logs, no new tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/397601
Original-Revision: 627d99dc752cb8e9cf531eb7f4dd53cce4ab3a4e
GitOrigin-RevId: 1aa34a73285a204405c091d482ddce08549b7426
Change-Id: Ife2075e1b9c47990b39c9fcb182d81bf4a8f99d6