[roll] Roll fuchsia [fx][test] Detect color output with tput
Dart uses the TERM environment variable to determine whether color
output and other ANSI escape sequences are supported. However, this
detection matches only on "xterm" - tmux and screen set TERM to
"screen", which is not supported. This is tracked as
https://github.com/dart-lang/sdk/issues/31606.

To force fx test to emit colour output the terminal detection is pushed
to the Bash wrapper script so that tput can be used instead.

Original-Fixed: 53267
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/395381
Original-Revision: 84912b5446b1fd82ede0b65d039ef90299837225
GitOrigin-RevId: 1946e73b94ca755d8eeb62c1957ce2639a591c6b
Change-Id: I4c8fb9f7dcc9007246b2a8300aa6526eebd039d5