[roll] Roll fuchsia [kernel][object] Ensure enumeration progress
This reworks the GetVmAspaceMaps and GetVmAspaceVmos to avoid
a race between performing a SoftFault and using the faulted in page.
Due to EnumerateChildren restarting from the beginning, for aspaces
with many mappings the faulted in page could get evicted before the
enumeration got back to the point of copying to it. As a result the
syscall could potentially spend many iterations making no progress.

By not considering the portion of the aspace that had already been
enumerated we can instead perform the fault+copy at the same time,
and then 'resume' the enumeration where we left of. Although the
enumeration may have to be re-walked several times, progress even
in the worst case is guaranteed.

Most of this commit ends up unifying the VmMapBuilder and
AspaceVmoEnumerator classes to share all the common logic instead
of continuing to duplicate it.

Changes to the ProcessGetInfoTest are simply undoing the error
code status changes that happend when the use of SoftFault was
introduced (If12c25f94c0d7e008f8fbdeefc007bc576d8082b)

Test: Correctness is covered by existing tests.

Original-Bug: 53685

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/397515
Original-Revision: 78ad4a86084d1e91d961a420bcaf26ee8dce8204
GitOrigin-RevId: a3ef605e9078d4ec6e0d72e4d52c15bee97fdc94
Change-Id: Ib0546c8944c7e5e6db9892d823e528d454b40ad9