[roll] Roll fuchsia [toolchain][rust] Fix compiler errors for upcoming toolchain roll
These are compiler warnings that are considered errors, and are related
to capturing variables within async move blocks. They point out
where a captured variable is unused and point to possible bugs in
the two places they were used.

The fix here is conservative -- the goal is to preserve semantics
of the existing flow. Owners are encouraged to investigate whether
there are bugs in business logic or test logic.

Original-Bug: 53510
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/397833
Original-Revision: 1fc64dda92d3d14ff9dcf104891cbe4aef83942e
GitOrigin-RevId: 980702dfebf78b238e57302e2b1561a231f54a01
Change-Id: If50ed4a88e2cf27e898acfa108ade3c925422de0