[roll] Roll fuchsia [dhcpv6] Ask for address while starting client
This offloads address selection to the caller and the client can simply
assume the address is available while starting. It is a better
separation of concerns since the management layer controlling the
client can simply use the information available to it and the client
doesn't have to query the netstack at all. This also gives the
management layer the option to start the client on non-default ports.

Original-Bug: 51074
Test: fx test dhcpv6-tests dhcpv6-client-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/397009
Original-Revision: cc326d3be91d9c82ebf684f8b77fe1cb6f21b23c
GitOrigin-RevId: cae6c0abd0cd9dd1aec70044afd5c77687979022
Change-Id: I59321aaf465b5d2acff4e000338b05083c8e6df8