[roll] Roll fuchsia [zxcrypt] Migrate to DdkUnbindNew.
To reply to unbind, UnbindTxn.Reply() is called rather than
DeviceRemoveDeprecated. Is is required to call Reply() or
save the UnbindTxn to be replied to later.

It is also guaranteed that DdkUnbindNew will only ever be called
once for a device added via DdkAdd. This means we can make the
following changes:

* Device::DdkUnbindNew
Since the deviceManager always calls Device::Init() before DdkAdd(),
and DdkUnbindNew is only called once, we do not need to check |active_|.

* DeviceManager::DdkUnbindNew
The device should not able to be in the kBinding state, as the state is
set to kSealed following DdkAdd while holding |mtx_|.
Since DdkUnbindNew is only called once for a successfully added device,
it should also not be in kUnbinding or kRemoved state.

TEST= runtests -t zxcrypt-test

BUG=37831

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/397813
Original-Revision: 223ee7a217ee82a14e3cde2333efbae8c00fa6ae
GitOrigin-RevId: ad3658ece07035f33a0abfdff0c790273cf0b569
Change-Id: If55bd1f34ebf0cc14f5655d0a6e9dbca5b32a773