[roll] Roll fuchsia [mediaplayer] fix end-of-stream detection in audio
In mediaplayer, the audio renderer (the wrapper for the FIDL
AudioRenderer) uses logic that compares the PTS of the packet marked
end-of-stream with the PTS of the packet last retired from the FIDL
AudioRenderer. That comparison could erroneously detect end-of-stream
when:

1) the end-of-stream packet has been encountered by the internal
   renderer, and
2) the FIDL AudioRenderer hasn't retired any packets.

In this case, the 'last_rendered_pts_' (in the Renderer base class)
is set to NoPts, which is a large number. The logic was not
distinguishing this case from last_rendered_pts_ having a large
'valid' value. This CL fixes that logic. It also resets
last_rendered_pts_ when the audio renderer is flushed.

These conditions are unlikely in normal operation, however, they can
occur in CQ when the fake AudioRenderer isn't getting attention from
the scheduler, and the entirety of the audio content fits into the
payload VMO used for audio. This is sometimes the case with the
PlayWav test that's flaking. It seems to be happening more in ASAN
builds, which may be due to the added CPU load of ASAN.

In the process of adding a new test for this case, issues with the
existing tests were discovered. Specifically, tests that retain
packets in the fake audio renderer and then release them were not
retaining packets at all. Furthermore, they were checking that
end-of-stream was false before the test really started, so that
check would always succeed. The issue was that the action of
retiring packets (SetRetainPackets(false)) and the EXPECT_FALSE
for end-of-stream were executed immediately rather than being
added to the command queue.

Original-Fixed: 53582

Test: fx test mediaplayer_tests
      Added a new integration test for this case.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/397716
Original-Revision: 88b0d434ac0a50d6f32b76bf339d1874e625d41f
GitOrigin-RevId: 83b8659b97a5e169557b555a65d4b47e5369a8e4
Change-Id: Ic8ea4cb7d0115045079089819475cb8857dc77cc