[roll] Roll fuchsia [brcmfmac] Add sim support for ARP offload
Implement support for simulation of ARP offloading. Note that at
this time we don't actually generate ARP frames (as the firmware
does), but we do inhibit the relaying of frames to the driver when
appropriate. For now, this is sufficient for testing the cases we
care about.

Original-Bug: 52305
Test: Included (arp_test.cc)
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/396778
Original-Revision: 774f0ee4c0f8618db912ca063cc5664db49cbb56
GitOrigin-RevId: efbc660a3048b8f3e70605cf292e12deceb8a759
Change-Id: I6b6210a79ec970c9c7fcea0aaa27157a72d0da27