[roll] Roll fuchsia [sshutil] Use lower log levels
A lot of these logs show up pretty frequently and aren't actually a big
deal, especially for the OTA tests, which intentionally reboot the
device. So I'm lowering the log level for a bunch of these logs to make
them less frightening. If something actually unexpected happens, it
should be the calling code's responsibility to log the error at a higher
severity.

I also added an 's' to the end of the duration in the keepalive timeout
log to be clear that the duration is in seconds.

Original-Bug: 48042
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/398259
Original-Revision: 0d9ad34e2de47c6d454140412efeac9fefbcd8c4
GitOrigin-RevId: 5fef6d516a93eff7624b7446a3e69a0bc7e74d08
Change-Id: I4674f39419ed77332f0c2555149b33506bc2dda0