[roll] Roll fuchsia [brcmfmac] Don't enable ARP offload in non-promisc
In set_multicast_list, we were setting ARP offloading to the
inverse of promiscuous mode. This ends up overriding the default
interface setting, which can result in ARP transactions being
dropped (most notably, in AP mode).

To resolve this, we now only override the default setting when
promiscuous mode is enabled (which at the moment is never),
which is more consistent with the intent of the original logic.

As a side effect, replace the flags bitfield of the net_device
struct with a boolean since we were only using a single field
(IFF_UP).

Original-Bug: 52305
Test: arp_test.cc
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/396780
Original-Revision: 5998dd3780cb4d74cb2f538890f02727f2403895
GitOrigin-RevId: ea323b4b6e71de872a64c7eea29eeb7ced7237cf
Change-Id: I2320828d2eec3bcf94381e8e6d9ce101eb92a1d3