[roll] Roll fuchsia [scripts][fuzzing] Streamline CLI testing
This CL reduces how much FakeCLI has to fake by providing a more
flexible way to "hijack" stdio and replace it with StringIOs when
testing. This allows more of the "real" CLI code to be tested, and will
enable better end-to-end testing.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/386239
Original-Revision: c6b90163710edfbcbdf02fe3fe7cfa140efb9337
GitOrigin-RevId: 513ea3fa0712eee09609fdcbbb4cc87351022288
Change-Id: I5cd78dde378fecfbef125c6ed1cd65c4e0c0a778