[roll] Roll fuchsia [modular][flake] Fix use-after-free in StopStoryCall
StoryProviderImpl, which owns `story_runtime_containers_`, may be
torn down before the async task in StopStoryCall that uses it
is scheduled.

The extra async task to destroy the StoryRuntimeController does not
seem to be necessary. None of the callbacks for StopStoryCall use
the StoryRuntimeController, and it's never called with the objects
that it's destroying in the stack (i.e. only called directly through
StoryProviderImpl).

Multiply: session_shell_test (fuchsia): 10
Test: fx test "fuchsia-pkg://fuchsia.com/modular_integration_tests#meta/session_shell_test.cmx" -- --gtest_filter=SessionShellTest.StoryControllerAnnotate --gtest_repeat=1000
Original-Bug: 53597

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/396957
Original-Revision: bb6c10831bedf96c861318b5daaec2ad622ebabc
GitOrigin-RevId: 5c42f5a735d0f1e322219b388e389ff904a1abed
Change-Id: Ib0e445c071a249feda351e8e55a786613628268d