[roll] Roll fuchsia [isolated-ota] Emulate actual paver behavior
When ABR isn't supported, the paver service closes the boot manager
channel with an epitaph indicating it is not supported.  Due to
limitations in the go bindings, the go system updater had to interpret
multiple error conditionas as ABR not being supported, but the Rust
implementation is able to read/parse the epitaph.  This change modifies
the isolated ota tests to use the epitaph so it is compatible with the
Rust system updater.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/398268
Original-Revision: 7296f15b7bb3ed35bc0ee67a7a58c789551375c8
GitOrigin-RevId: 0d220875e7caa37a66e2e303649f8b639dfd40eb
Change-Id: Id29bf30f6316226044e0dd139c0560533dd02165