[roll] Roll fuchsia [codec_impl] Fetch Auxiliary Buffers
An auxiliary buffer collection is a set of secondary buffers that can be
used to provide additional context between two pipeline nodes. The
current primary use case for these buffers is for decryptors to copy the
clear input data through to the decoder. Typically, when the decryptor
is decrypting into secure buffers, the clear input data just gets copied
through to a secure buffer, so it is unable to be read from an
unprotected context. This data typically includes metadata information
that is useful to the decoder's parser that might be run in an
unprotected context.

Based on the CodecAdapter implementation, CodecImpl will now fetch
auxiliary buffers if necessary and provide mapped access to them via the
CodecBuffers that are shared with the CodecAdapters.

TEST: fx test codec_impl_unit_tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/396230
Original-Revision: 8e1c22a91198bb41700fcf9fd33b63ac2cbe7712
GitOrigin-RevId: 9f46d4658d3e9e33b9c72847105e58b48ec4601f
Change-Id: I289d6bf0783616525601209d18cdea8be7cb0c8e