[roll] Roll fuchsia [artifactory] Set GCS timeout 8 minutes and chunksize to 100 MB
The fundamental problem we are facing is the number of round trips between
the host and the cloud. Lets say there are files of ~5GB. With a chunk size
of 8 MB and average roundtrip call taking about 500ms-1000ms, we would
take 320 seconds(0.5*5000/8)-620 seconds to upload that 5GB file.
Sometimes, the time will exceed the limit causing about ~10% of failed
builds. By increasing the chunk size, we are reducing the number of
round trips. This reduced the average time to 120-240 seconds.
(Note that it will not be linear considering that per call latency will
go up with increase in chunk size but the number of roundtrips will go down).

Testing: Validated that the uploads (of ~5GB) failed (http://shortn/_8XfWaADEEl)
before the change and the same build succeeded (http://shortn/_iE3jesq9dB)
after the change.

Original-Bug: 52317
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/396479
Original-Revision: a23b49e0a5cbaf014d2f348d2e56ff46a7d78cd8
GitOrigin-RevId: c684ea401d3f82e0aa2376f0c7a738a9ddccedea
Change-Id: Ibea4d4d00d730655d157fd488107de54f75182b1