[roll] Roll fuchsia [scripts][fuzzing] Rename CLI to Host
As the CLI class has grown, it has taken on more and more of the
platform-specific code, and the class formerly known as "Host" has
become more and more focused on the details of the build
environment.

This CL is the second part of a renaming to reflect that shift; it
renames "CommandLineInterface" to "Host".

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/386241
Original-Revision: 9186ccbbfc52faae9707b9005b7cb06574e707ca
GitOrigin-RevId: 912092b66528b5f35fcf64527041631c5a9e9cb6
Change-Id: Ic41dc708129a28802dc3dd1b3d53c56122f25d3c