[roll] Roll fuchsia [wlan][policy] Record connect failures
This change creates a function in SavedNetworksManager for recording
failures when attempting to connect to the network. The NetworkConfig
already contained a field for remembering failures to connect, but this
CL renames them to be more clear. There is no way to use the BSSID that
was in the existing field so it is removed for now.

Failures (and successes) are recorded in mod.rs which handles API calls
(without using the state machine currently) and inside of the client
state machine.

Original-Bug: 53545
Test: added a unit test and modified existing tests:
	fx test wlancfg-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/397800
Original-Revision: 8491eb054c8e049df6e71b1eb662894e871f54da
GitOrigin-RevId: 10ea1ebd8c432da9bc26e4553db21084662bd18c
Change-Id: Ic46effd3b6c30f1f77ed85615ecb5e71ff4eef8f