[roll] Roll fuchsia [kms-stateless] Add fallback for finding tee.
The old logic using fdio_watcher gets a ZX_ERR_NOT_SUPPORTED when run
inside of a component. I am assuming this has to do with the way devices
are presented to the sandbox.

This logic isn't critical so we just fallback to a fairly dumb polling
approach if we see that the watcher isn't supported.

Test: tested using a new component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/397793
Original-Revision: 9cccf17ba5e792a951f16ae2e855502c9d1654b7
GitOrigin-RevId: c33859df70fc05b41d2545731f5b72581904ad0e
Change-Id: Ib7effdf5b7f098d73471791f9908c846dc8444cb