[roll] Roll fuchsia [cobalt] Rename cobalt_testapp_no_environment to cobalt_testapp.
The name "_no_environment" no longer has any meaning because
no version of the Cobalt testapp uses the version of the
Cobalt service from the environment like it used to. All
versions of the testapp do the same thing: they start their own
instance of the Cobalt service.

We now have three versions of the cobalt_testapp

(a) cobalt_testapp - uses the network, is only run manually
and not on the bots.
(b) cobalt_testapp_no_network - does not use the network and
is run as part of Fuchsia CQ
(c) cobalt_testapp_for_prober_do_not_run_manually - this is
idential to cobalt_testapp except that it uses a special
Cobalt project that allows us to keep track of the data on
the server side and thus it acts as a prober. This is run
in some special Cobalt CI bots called the Cobalt probers.

Additionally in this CL:
- We moved the cobalt_testapp (formerly known as _no_environment)
out of the cobalt_tests package and put it in its own new
package called cobalt-manual-tests. This package uses the
new-style GN templates fuchsia_component and fuchsia_test_package.
- We need a way to prevent the test from being run in CQ.
In the previous GN configuration this was done via setting
disabled=true. The new GN configuration does not suport this.
Furthermore if we set disabled=true then the test cannot be
run using fx test. So in the new configuration we take a new
strategy and use environments = []

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/398373
Original-Revision: c5e5ec031379574ce2b2048429fe4adb2c1f748e
GitOrigin-RevId: c6b41104575160f28cc41873b0592b2cc5edd8dc
Change-Id: I71a60b14a8cbc6a149f62a4748b664960ccc1573