[roll] Roll fuchsia [bt] Remove ununsed RFCOMM code
As part of the plan to move RFCOMM support to the other side of the
Profile API, this code is being removed.

It was not in use anyuhere, and bugs in the code are not going to be
addressed before the transition.

Original-Bug: 49073
Original-Fixed: 50629
Original-Fixed: 37806

Test: only removed code, bt-host-unittests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/398534
Original-Revision: eed5c7ed66ca8157b1dee0a8520f6f38a8f5eedd
GitOrigin-RevId: 14dec5f7a24281454511168ddd53976c9022844a
Change-Id: I4985982c5e1a9b38a152f3fa25e6d4fbc5b1392d