[roll] Roll fuchsia [components] Implement EventStream & CapabilityRequested
This CL introduces two features to enable connection attribution in v2
components:

1. CapabilityRequested event:

The CapabilityRequested event is a "move-only" asynchronous event that
enables client attribution of connections.

It holds on to the server end of a channel. It has a follow key security
properties:

a. Only the source of the capability will receive events corresponding to
   capability itis servicing. If it recieves the event then the framework
   will not attempt to connect to the capability from the outgoing
   directory.

b. The CapabilityRequested event cannot be offered and it cannot be used
   from the realm.

c. Only one event stream can receive the event. Any other event stream
   within the same source will receive an error event.

2. Static EventStream:

A static EventStream is an EventStream at is specified in a component's
manifest file and subscribes to events as soon as the component is
resolved. This is used to trigger the start of a component upon receipt
of a subscribed event.

This CL includes unit and integration tests that cover the
nuances mentioned above.

Original-Bug: 51179
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/395903
Original-Revision: 53ed8569e6b9f370be9c9ebdd8427837d4b8204e
GitOrigin-RevId: ee826c3250b507cb40a16f56262019a5357d1a5f
Change-Id: Id9dc1a9c16572dc88c78970e4de8f1db70efbd6c