[roll] Roll fuchsia [kernel] Ensure RootJobObserver created after the root job has children.
Currently, the RootJobObserver --- which restarts the system if the root
job is ever without children --- is created before the root job actually
has any children.

We get away with this for the moment for two reasons:

  * We don't check for children in the RootJobObserver's OnInitialize
    callback;

  * The root job doesn't happen to set/clear any signals prior to its
    first child process being created, so the OnStateChange callback
    doesn't fire prior to the root job's children being created.

Upcoming changes to the observer API will prevent the
OnInitialize/OnStateChange checks being different.

This CL delays creation of the RootJobObserver until after the root
job's child process is created. We keep ownership of the RootJobObserver
with the Executor, but create a new function to allow userboot.cc to

Original-Bug: 51463
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/395761
Original-Revision: c59b32e2487b48a3b8e3698ad7543c6cf36b1367
GitOrigin-RevId: 51b9a291dc84a8504f97b1c1506707e8ef013b02
Change-Id: I70817bd1a62251c4f04487aecfff3327ddc665b7