[roll] Roll fuchsia [llcpp] Remove move constructor from OwnedSyncCallBase
This calls into fidl::Linearize, which I'm trying to remove.

It is also only used in tests and is expensive to call outside of tests.

yifeit@ mentioned that the move constructor is needed for C++14 support,
though we need further refactoring for that anyways.

Test: mostly removed tests, but fx test llcpp_types_tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/398123
Original-Revision: 2294a32e9d982cb223e976f817d3ecdc04461136
GitOrigin-RevId: b0a9964b3829fa6289a5d687db62c81e7fd5d98e
Change-Id: I2c9cd0a6f2cc7b24840f75463b113fc4fa41e96e