[roll] Roll fuchsia [tools][netboot] Don't bind to debugPort when sending netboot query.
In order to discover a device, we bind to a netboot client port and send
a query to get the address corresponding to a specific nodename. One of
the ports in this range is the debug log port used by loglistener. This
occasionally causes loglistener to fail with "address already in use" if
another bot on that machine has already binded to that port.

Original-Bug: 53262
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/398543
Original-Revision: 80af43292a4450f652ee2dab64a29b4797848144
GitOrigin-RevId: ea3420c3be63776f13b4e0ed545478c4db42078b
Change-Id: I66f7603dd97a53c30558d899b60d4437dd0ba399