[roll] Roll fuchsia [bt][a2dp-sink] Respond to stale absolute volume requests
The Volume Relay for A2DP Sink stores the OnVolumeChanged responder and
only updates when the system volume changes. However, it is possible
that the peer that issues the OnVolumeChanged disconnects before
receiving the response. Consequently, the stored responder will continue
to exist in the Volume Relay. However, when a new peer connects to AVRCP
and A2DP and makes an OnVolumeChanged request, the FIDL stream gets
clogged, because the previous (stale) responder still hasn't been
resolved and is dropped.

In practice, the hanging-get pattern requires only one outstanding
request at a time, so the onus is on the client to not further requests
until the first has been resolved.
Regardless, it makes sense for the server-side to have an appropriate
policy, and it is reasonable to store all outstanding responders and
respond to all when the volume changes.

In production, this bug manifested when a remote device (i.e phone) was
streaming music, disconnected (while media was playing), forgot the
device, and attempted to repair + start streaming media. In this case,
media streaming was OK but volume control crashed for the reasons
mentioned above.

Added TODO to relevant sections pertaining to larger scale AVRCP fix
around hanging-get semantics.

Original-Bug: 53997
Test: Added test_volume_changes_with_peer_disconnection,
test_volume_changes_with_multiple_peers

Multiply: bt-a2dp-sink-tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/398395
Original-Revision: 7a91992a4cbca5b19667234fcd5281cad7780949
GitOrigin-RevId: f39a2dd5f2b22c01f2ee3abb5bddfb4bf0d6317c
Change-Id: Ia5a41e281944bb308ab527f1697bbd1ce6914625