[roll] Roll fuchsia [fdf] Fix logs in device recycle
During device destroy, fbl_recycle() method is called and within
this method, creation of Ref_Ptr of the device is not allowed.
See comments in RefCountedBase::AddRef() for reasoning. As this
is used in logs, the assert in AddRef() gets triggered. Changing
mkdevpath to fix this.

Test: fx test driver_host_tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/398502
Original-Revision: 759828e3dbd2996887ab9f4ca071377d3a5b3c5c
GitOrigin-RevId: 4bf18a083cc31f43dd043a99ae5bba4a4640c109
Change-Id: I151493d7cbdf46834e1a5ebd884da4ba6c175be3