[roll] Roll fuchsia [components] De-conflate CapabilityRequested event target and source
The previous CL really conflates target and source. Partly, this is a
problem with the code itself because an EventDispatcher also has a
target which isn't the same target as the target of the event.

EventDispatcher target is where the event is getting sent.
The target of an event is the component the event is applicable to.

Capabilities in component manager also have the notion of source and
target.

CapabilityRouted sets the target of the event as the target (the client)
of the capability.

CapabilityRequested was setting the source (the server) as the target of
the event. This is inconsistent with CapabilityRouted and not particular
useful for log attribution. We want to know the client (target) not the
server as this event is only sent to the server of the capability.

This CL fixes this by introducing a source_moniker to the internal
CapabilityRequested event. This CL also sets the descriptor of the event
to the client (target) of the capability.

Original-Bug: 51179
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/398833
Original-Revision: 6a640d4aee49ee0085470092c2b78e6185bbe916
GitOrigin-RevId: 9a8ae349894d08c9b9ee28b2350a84413376970d
Change-Id: I7a9d774f021dd3ff35e32b90f9f407a68a9de8cb