[roll] Roll fuchsia [brcmfmac] Clean connecting bit correctly
brcmfmac driver skip cleaning the BRCMF_VIF_STATUS_CONNECTING bit when
SME issues an assoc_req while another connecting is going on in driver,
the skip happens because SME will send a deauth_req when we trying to
connect to the same ssid again, which will trigger
brcmf_cfg80211_disconnect for this ssid in driver, but if the input
peer_sta_address is not the same as the one we saved in
brcmf_80211_profile(it will only be modified when connection success),
this function will return a failure and skip cleaning the connection
status bits.
Now it will not return failure and will print out a debug log instead.

Original-Bug: 51408
Test: Manually test with wlan tool using two shells for one device, the
      test flow is:
      1) Hit "wlan connect" in shell A
      2) Hit "wlan connect" to the same ssid when the join scan of
      connect finished but the connection result is not returned yet.
      3) deauth_req triggered and the second connect will return a
      success while the first will be returned as "Connecting was
      canceled or superseded by another command" in shell.
      Automatic test added in timeout_test.cc in fxr/391953.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/389216
Original-Revision: 3d51eacc36ff7d892023a60709060e3fc1fa828e
GitOrigin-RevId: 5d66fb261d92d4807ce753846ddf6b6e0f5c977d
Change-Id: I3489818415274aef8529c2cc2d84284952eeec90