[roll] Roll fuchsia [brcmfmac] Workaround for zero rssi signalling data.
The bug in firmware causes numerous zero rssi readings in
signalling data. This is due to the fact that firmware only
indicates the rssi in the first packet in an ampdu. The workaround
is to save the last non-zero value and uses it as long as the
signalled value is zero.

Test: Added a SIM test that tests driver's handling of 0 RSSI
values in signalling data.
Manually tested and do not see any zero indications in
the rssi stats (wlan iface stats).

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/390627
Original-Revision: 1df14fd9023a09dd6c79b9fe3c6eabe99a83d96d
GitOrigin-RevId: d35a76e1da49b0c4b5fee257e24239a9a5d44cab
Change-Id: Iddee0be0dc70a90ea612e311844b5493d8763204