[roll] Roll fuchsia [fidl] Fix documentation bug
The comments for AudioRenderer.Play say that reference_time is relative
to the system clock. However, the implementation makes this time
relative to the provided reference clock (when a reference clock is not
explicitly specified, it defaults to the system clock).

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/398565
Original-Revision: cb39aefe88ac7f9075fb5e8e211411a3783ac26d
GitOrigin-RevId: fac2d922d472238b641a202896ed3611bf719bf1
Change-Id: I48c4766bb1935371b57e093cabdb02b3d03f587a