[roll] Roll fuchsia [async] Add async::WaitOnce type variant.
Unlike async::Wait, async::WaitOnce moves the handler lambda to the
stack prior to calling it. This enables the handler to safely delete
the async::WaitOnce that triggered it.

Because of this, WaitOnce also uses a different signature for Begin()
to accept a new handler each time.

TEST=fx test -o async-test

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/396270
Original-Revision: b528e3c1ec0e008bb68376c75c71d6bf71817bbe
GitOrigin-RevId: b99d112d39c64840189d22fa0507e0e51e75394d
Change-Id: I9fd03393e70cd9bfa3b11bede551d3c1f9cf9b0c