[roll] Roll fuchsia [scrutiny] Add command line argument parsing and forward request body.
Scrutiny in the long-term will need support for command line parsing.
This change introduces the ability to configure the port that scrutiny
runs on. It allows for the user to provide -p or --port option to set a
port. This change also forwards the request body to the dispatcher and
subsequent DataControllers.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/397497
Original-Revision: 4b16580e80765c0300a74a63facf4b467e016eb5
GitOrigin-RevId: e83fdcafe59068813a071cb9e7039e18a8279d07
Change-Id: I694842eb6d22e37587bcbeaa820560d0206a6ef1