[roll] Roll cobalt [cobalt 1.1 privacy] Add privacy encoding for UNIQUE_DEVICE_COUNTS.
The PrivacyEncoder will be called by the ObservationGenerator in a follow-up CL.
Also adds a PrivateIndexObservation type to observation.proto. This is the
type of observation that the PrivacyEncoder will produce for all reports
with a nontrivial privacy level.

Co-authored-by: pesk@google.com
Co-authored-by: azani@google.com

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/cobalt/+/397316
Original-Revision: ee96afb818cd2a04660ee6b1b4d6dea5f3ad2944
GitOrigin-RevId: 29db5240c7e71992a0276b958380c5b9eb71581f
Change-Id: If8f0133ba6e273a7bfa417396788a9040c7ce500