[roll] Roll fuchsia [audio_core] Wire device clock into the pipelines
With this CL, the device clock is passed up into the input
and output pipelines. A future CL will add reconciliation
between the client clock and the device clock at the place
where they meet, which will be the top-level mixer (closest
to the client). With this CL, the MixStage is directly
passed a reference clock for its destination stream, and all
other parties "ask their neighbor" for the clock -- a chain
that eventually leads to a device, MixStream or PacketQueue.

Original-Fixed: 46629
Test: new OutputPipeline unittest

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/387459
Original-Revision: 583dfa248183e1d7358aa5fa9e84e8c563061970
GitOrigin-RevId: 1e65d3f2dfde922d2f85caee3358c3702984bdf7
Change-Id: I077e189b4ddef67cdbe5082aaf5ee90de583ffc2