[roll] Roll fuchsia [zircon][inet6] Fix a buffer read overrun in ip6toa
This was located by ASAN during testing and ended up as flake-failing
an unrelated change that I was attempting to land.  ASAN provided a
pretty solid report, and the code itself allowed for a rather simple
repro.  Between the two of them, it was not too difficult to find the
actual mistake.

Restructure the code so that it is (hopefully) a bit easier to read
and understand, and a bit harder to make mistakes with.  Also add some
tests as there did not seem to be any before this.

Tests: new unit tests.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/399101
Original-Revision: 8194e795ad65d4b3944530b2b8f44e78479b3761
GitOrigin-RevId: af5e781f830a85160c239bcceb708ec30be18450
Change-Id: Idcce8a0458995b1f5125e98b7cd7c51fa80d6333