[roll] Roll fuchsia [sshutil] Make ConnectionError a struct
Previously it was a sentinel error:

  var ConnectionError = errors.New("SSH connection error")

However, a sentinel error can't wrap another arbitrary error, so it's
impossible to idiomatically determine the "cause" of a ConnectionError
using errors.Is or errors.As. This change makes it possible for
ConnectionError to wrap another error, which will be particularly useful
for making changes to the sshutil library to wrap errors in
ConnectionErrors without obscuring the underlying error that callers
sometimes check for.

Original-Bug: 48042
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/399624
Original-Revision: df3404678d96068fe9a37a08375a18448454669a
GitOrigin-RevId: 8c5e85097d09cee70102dbe1f8fde4fff0072beb
Change-Id: I8035744aac3f0d18c23442d134beaa1dcfc59766