[roll] Roll fuchsia [tools][affected_builders] Add GN commands to tool
This adds a path to a GN binary as a parameter to affected_builders.

It also adds a function that takes the provided gn binary and runs `gn
analyze` with it.

The output of the analyze call is returned to the caller.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/399616
Original-Revision: 584a6c37bbf056b19dae899c84eaee013a7019a1
GitOrigin-RevId: 8478d3f478659a4fb08f85ecd3e7609e7c1e2810
Change-Id: I10a94277d9b2473fdd134c4992918bc16ebc491d