[roll] Roll fuchsia [hypervisor][arm64] Correctly restore active LRs
If an LR is not inactive, make sure that we correctly restore it when we
return to the guest.

This involves checking whether an interrupt in any state is in an LR.
This ensures that active interrupts are not added twice, and that
pending interrupts that have been popped from the tracker but not
serviced by the GIC are not lost.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/399236
Original-Revision: e874f07efd631dba20718b9a09f962fa65668c11
GitOrigin-RevId: 30f2d4bde28df0d4cec467924c4f58a6ac1db938
Change-Id: If6facfcc93e9989568d6f5e85f9c4851b00c0c6a