[roll] Roll fuchsia [blobfs] Fix off-by-one error in Blob::SpaceAllocate
Blob::SpaceAllocate is reserving an extra node for each blob
in memory. This reduces the number of available nodes according
to the BlobFS process.

Note that this bug only affects reservations made against
the node map that is stored in-memory. It does not affect
the actual number of nodes that are allocated on disk. Hence
this bug does not appear if the BlobFS process is restarted,
since the in-memory state will be correctly initialized from disk.

Created new test /system/test/fs/blobfs-large-integration-max-reservation-test
which exercises this bug.

Original-Bug: 54001
Test: Ran blobfs unit, integration and large integration tests
      (fx shell runtests -n blobfs-integration-test)
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/397999
Original-Revision: f90360b7683b1e369d231b9d7eb47a6b84a798a7
GitOrigin-RevId: 7fca8009bf27e1d03d9ff37e85b1b0205164ee21
Change-Id: I4b07df88cc360225718e2ba1f41e966ca4bcc370