[roll] Roll fuchsia [kernel][thread] Have Thread own ThreadDispatcher
Thread should hold a RefPtr to ThreadDispatcher
instead of doing manual refcounting in
ThreadDispatcher. This requires some changes
to the exit path as Thread should be in charge
of the exit flow.

As a bonus, the DPC in ThreadDispatcher can be
folded into the DPC for Thread. It is still
necessary to create a DPC to cleanup the Thread
because a Thread cannot free itself while it's
running. We only need a single DPC to take
care of Thread and ThreadDispatcher together.

Original-Fixed: 48779
TEST: runtests /system/test/core
      k ut all

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/374934
Original-Revision: 68394eba6ca7902bcfa1f8083cc3b718be1dbf18
GitOrigin-RevId: 08c7198ab22dc4524943ec6bd86cd0cbdb4b012f
Change-Id: I7edd038cfd6f873f8667e40dd585d9fc6fe3abd2