[roll] Roll fuchsia [components] Only dispatch CapabilityRouted for namespace capabilities
Dispatching the ELF runner capability prior to the Started event was
really confusing and causing people to spend a lot more time debugging
tests than they needed to.

This CL ensures that the CapabilityRouted event is only dispatched
in BlackBoxTests for namespace capabilities.

In the future, if we wish to be able to mock or interpose runners,
we can introduce, a new, separate event type.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/398965
Original-Revision: be6e02d9504c5781c0e243756319bfc44499bc0a
GitOrigin-RevId: 2f70d5b7fea04366e26d0f04384bf67b6d7eeed0
Change-Id: I31e56db1a9c8f0ab2b5dca8163ba7ecd682b2367