[roll] Roll fuchsia [rust] Add a Task abstraction to fuchsia-async
This is an initial change to get in place an API that we like. Later
it's likely that we'll reimplement these ideas and fuse them with the
implementation in Executor, but that's not needed just yet.

API is a reduced version of the smol::Task abstraction.

Two spawners are provided: spawn & local - for task spawning on the
current executor and on the current thread.

Tasks cancel work upon being dropped, unless they were 'detached' in
which case they continue executing in the background.

Tasks are themselves futures, and so can be await-ed on. This forms a
join() functionality.

Tasks are typed, and that type is returned from await.

MULTIPLY: fuchsia-async-tests: 100

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/397701
Original-Revision: 6a90ea5a095341f93424105cb6b98b537d8ff805
GitOrigin-RevId: 1757b85bf41c4c696301beed107cba2125b8c26f
Change-Id: Ibeabe479adaa560060298e6379a4a63ba340c467