[roll] Roll fuchsia [power_manager] Clean up shutdown concepts
With the new shutdown path well into implementation, the power_manager
needs some consolidation and clean up with regards to its handling of
system shutdown. This change attempts to capture the concept of system
shutdown as it exists today, and moves to a cleaner internal
organization to support it.

Test: fx test power_manager_tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/398270
Original-Revision: d5203c7be42b4256b398ab92df075de0b46ed965
GitOrigin-RevId: 979d9a82169ec3f37090b1b910eaf425cd138176
Change-Id: I660fe1ca13b6b388323e4935fd15955a1a62d21c