[roll] Roll fuchsia [modular] Move some of per-agent init to AgentContextImpl constructor
In preparation to make session shells an agent, move init of all
services provided to agents, and launching of the agent component, to the constructor of AgentContextImpl. Session shells and regular
agents differ in what services are provided to them, and who has
responsibility for launching the sub-component, and this change
makes it easier to share the initialization code that they have in
common, and not the rest.

Original-Bug: 53687
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/399306
Original-Revision: 9a14ba1ef0be364473321a24cfc7c81d1cb567e1
GitOrigin-RevId: 7105dd5547bad45e92cfa5327e9381f13e7c607c
Change-Id: I46292f5a842ad251c56aafe071cffc7eafcac541