[roll] Roll fuchsia [kernel] Add load_balancer class, load shed thresh
Create a load_balancer class. It is designed to be run periodically to
take a global look at the state of the system and make decisions to balance it.

The initial decision is to look at the load on all cpus and calculate a
threshold at which they should start load shedding.

Background: go/zir-loadshed
Test: Only unit test for now.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/362020
Original-Revision: 54f79135d140db56ae1f68ecaa872d048c1d7ac9
GitOrigin-RevId: 0a98b6b35be300d763a88fda3f0b539059a4a6ce
Change-Id: I34b565763a7d53a7131a65141b9495577e478b81