[roll] Roll fuchsia [ldso] Minimize stack allocation for DT_REL processing in ld.so
ld.so itself gets relocated twice: once for "bootstrap" relocation
before other objects are loaded, and then a second time for "final"
relocation after all objects are loaded and final symbol resolution
can be done.  In DT_REL (SHT_REL) format, the addend for each
relocation record is stored directly in the relocated data field
itself, so the original value is clobbered by relocation.  Hence the
original addends from ld.so's own relocation records are clobbered
in place by bootstrap relocation.  To preserve the original addends
so they can be used for final relocation, ld.so copies the addend
values into a stack buffer clobbering them in bootstrap relocation.

Previously the stack buffer was allocated with enough slots for all
of ld.so's relocation records (except the special R_*_RELATIVE type
that is only processed in bootstrap relocation and need not be
processed a second time at all).  However, the common GOT/PLT types
that comprise the vast majority of ld.so's (i.e. libc.so's) own
relocations never use addends at all and so don't need slots in the
stack buffer.  This change fixes the logic for sizing the stack
buffer to match the logic for using it, which makes the size
actually needed much smaller.  The hard-coded upper limit on this
size is likewise adjusted down to a healthy breathing room above the
current actual requirement.  This limits the initial stack usage to
the range predicted by the program loading logic even when ld.so is
built using `-z rel` so it uses DT_REL encoding.

Original-Bug: 53078
Test: follow-on change to build with `-z rel` breaks everywhere without this, works with it
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/400021
Original-Revision: abe6b4d2762afe25526e0b6202e4b5ff4178f922
GitOrigin-RevId: 603a9eb31db8c2552c0eeddb866b422eb52b6ae2
Change-Id: Ib96145346353219ace99734332a9eec4e520cfa2