[roll] Roll fuchsia [libc] Separate some utility code for getifaddrs into a header
The implementation of getifaddrs is going to move to fdio, but much of
the address copying logic will remain the same. This CL places that
shared code into a header, and adds a BUILD.gn file to share it with
fdio.

Original-Fixed: 54214
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/400025
Original-Revision: 8eeaac245c66e6473f8260d659001bd61543dd0f
GitOrigin-RevId: 78ab1f74e7c6ae91d7bf24cf5353f24bc820a322
Change-Id: Iacb48247a3bc65469965c670cd1c1b5cc84b3afc