[roll] Roll fuchsia [fshost] Serve /data stats in inspect
This is currently done in the archivist, but this is a much better place
to do so now that we can expose inspect from fshost.

This removes the need of global-data in the archivist which will be done
in a follow-up once clients of this data have been updated to use the
data exposed here instead.

Original-Bug: 52951
Tested: fx shell runtests -n fshost-inspect-test-test

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/395360
Original-Revision: c79bca4d04386b37194d84d3962557589632662c
GitOrigin-RevId: 5cfb90038e2cd8a5d2269db4979e235135f2a173
Change-Id: Ifc67bfa551a6213dc1f263725b4d1c106afae169