[roll] Roll fuchsia [storage] Add the equivalent of RemoteBlockDevice for Rust.
Also fixed a bug in our storage driver that would drop responses if the
fifo became full. This isn't a problem with our C/C++ client because it
limits to a maximum of 8 in-flights, but the Rust implementation doesn't
have this limitation.

This does not support barriers or flush (yet).

Original-Bug: 53796
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/399393
Original-Revision: a92603d23d7c712d149fa53302731e1b22fc7e12
GitOrigin-RevId: 94ef3d41abac909bf36576ddc4334cdcd9ed874e
Change-Id: I7fbd3cbe07156983390ce42e9e2da90e50fe4bb9