[roll] Roll fuchsia [gn sdk] Adding CIPD packaging rule templates
This adds the rule templates to create a CIPD package manifest.
It also adds a template to create the .build-id directory containing
the GDB compatible symbols for the binaries contains in the Fuchsia
packages.

Using this approach to upload a CIPD packaged .build-id directory is
preferred over the tarball approach.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/399479
Original-Revision: fbf28945442ed9bbf323cfbd25d5d3cde00f8c19
GitOrigin-RevId: ec09850182a4bf62fbd8d49864a38ab120c0d19f
Change-Id: Ic8bb03c8aab25d777c4a1823299326f877d4ce9b