[roll] Roll fuchsia [IDK sysroot] Removing syscall FIDL files from headers list.
Currently the .fidl files in //zircon/vdso are being added
to the sysroot header file list with a relative path of ../../vdso/.
This breaks the Rust tar processing which does not allow ParentDir
relative paths in the tarball manifest.

This also creates an undocumented directory in the SDK at
sdk://arch/$cpu/vdso.

This change removes these files from the sysroot SDK atom.

Once the tooling is ready to be exposed for processing these
FIDL files, they should be added back as a fidl library atom and not
part of the sysroot SDK package.

The GN SDK generation script is also updated to accommodate the
new sysroot metadata.

Original-Bug: 54116

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/397364
Original-Revision: 1ba43120b7f6686d6817fc791f7ce9a0f32d8958
GitOrigin-RevId: 9e47ed88af564df2f441a9d248029fca85f1868d
Change-Id: I273417c1ece6327e13f71057faf8cb38024179b9