[roll] Roll fuchsia [fidl] Enable subset of {encoding,decoding,validating} tests on host
When working with the walker it hard to debug because there is a cycle
where you can't boot but the tests depend on you booting to run them.

This CL enables a subset of tests on host to avoid this cycle as much as
possible.

Test: fx build, then out/default/host_x64-asan/exe.unstripped/fidl

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/400356
Original-Revision: 1af75c63bf25f149346a71b0f64cb6dfe2f74dd2
GitOrigin-RevId: 8849fbc4271accd865d12f202e934a83c7bde29c
Change-Id: I96bcc25ffa81009dddbd64813e76116bfb1bd0c0