[roll] Roll fuchsia [scheduler] Call MigrateFn after updating state
In a recent change, we begun using Thread::LastCpuLocked() in migration
functions after CPU migration. This is used to update the CPU that the
thread is currently executing from, so that we can target IPIs
correctly.

However, the scheduler was not updating the thread's SchedulerState
until after migration, therefore causing us to IPI the wrong CPU. This
can result in a dramatic slow down of the virtualization integration
tests, especially on arm64.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/400253
Original-Revision: f782cc27002d7cf9d3567b3b96b511e202385137
GitOrigin-RevId: b14690cfda7dbeed989497b67b43010bccf4ae79
Change-Id: Ia95c1ec634f2e234d11f02e9e537f51020781eaf