[roll] Roll fuchsia [kernel] Add internal ZX_JOB_NO_CHILDREN signal to jobs.
Currently Job objects expose the ZX_JOB_NO_JOBS and ZX_JOB_NO_PROCESSES
signal to indicate that the job has no child jobs or child processes,
respectively. However, there is currently no way for watchers to be only
notified when the job has no children (i.e., no jobs and no processes)
at all.

The current work-around is to listen for one signal and, if it fires,
check for the other. If the second signal is not active, we wait on it
instead. We may have to keep going back and forth as new jobs/processes
are created or destroyed:

  zx_signals_t remaining_signals = ZX_JOB_NO_JOBS | ZX_JOB_NO_PROCESSES;
  do {
    zx_signals_t observed;
    zx_object_wait_one(target_job, remaining_signals,
                       ZX_TIME_INFINITE, &observed);
    remaining_signals = (ZX_JOB_NO_JOBS | ZX_JOB_NO_PROCESSES) & ~observed;
  } while (remaining_signals != 0);

This CL adds a ZX_JOB_NO_CHILDREN signal, which is active when the job
has no children at all.

Currently, we only expose the constant in-kernel, allowing the
RootJobObserver to use it in an upcoming CL.

fxb/53986 tracks either making this signal visible to userspace, or
providing a more general mechanism to userspace avoiding the need for it
altogether.

Original-Bug: 53986
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/395756
Original-Revision: 6c7fc50dcdf279e7676545ac26594b9718ba9b22
GitOrigin-RevId: aacd9a2a4a78b60ef16a9afaefdb7aff5f9e19e1
Change-Id: I7f8d6a186fbfec77a087797c44566bf7c110bf3f