[roll] Roll fuchsia [dma-buffer] Fix memory leak
This patch fixes a memory leak in dma-buffer.
The RAII wrapper for PMT does not unpin the PMT,
so the memory was not being properly freed.
This patch manually invokes unpin to unpin the pages
so that the physical memory can be freed when the buffer
is destroyed.

Test: fx test dma-buffer-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/393634
Original-Revision: 198cfc0b793a8967f50100f17436e202e8dc3bba
GitOrigin-RevId: 6d0ced17673f21c58be5191dbd2e5e0ab883b188
Change-Id: I00c3ae8bdfcb1a8812c2096935d8aa189f703b97