[roll] Roll fuchsia [perf] Migrate dart inspect benchmark away from tspec
Update dart_inspect_benchmarks_test not to rely on the tspec file
//topaz/tests/dart-inspect-benchmarks/basic_benchmarks.tspec.

Previously, this test would run for the shorter of 1000 iterations or
30 seconds. Since 1000 iterations has consistently been shorter than
30 seconds in CQ, update the test to always run for 1000 iterations.
This simplifies the code to launch the benchmark application using the
launch facade. Log the total time elapsed so this can be more easily
updated if necessary.

Since the number of iterations is now constant, add a check that each
expected event appears exactly 1000 times in the trace model as an
additional correctness check.

Use a map from event name to output test name to specify which events
to include in the benchmark, similar to the tspec file. As that file
does, this map excludes the following events from the dart benchmark:
* Init and get root
* Get root
* Get integer
* Dec integer
* Get double
* Dec double
* Add property
* Delete property
* Get string
* Get string long
* Get byte
* Get byte long
* Allocation init

The tspec file itself will be removed from the topaz repository in a
separate CL.

Test: fx run-e2e-tests dart_inspect_benchmarks_test

Original-Bug: 53025

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/399877
Original-Revision: 18d41e07f777806d9002f67020f29f728282bdc7
GitOrigin-RevId: d7e132f7acf6b4254584ea288344dbdab15ebbd6
Change-Id: Ia07150eb85818dd60d22b72732e6b812706833e2