[roll] Roll fuchsia [bt][sm] Improve ParseSdu safety
ValidPacketReader::ParseSdu had stored a `size_t` length value into a
`uint8_t` for a check. Fix this, as an overflow here would lead to the
method returning and logging an incorrect error. Also, clarify that the
method checks the packet size against its expected header size, not the
size of `Code` and change a DEBUG_ASSERT to an ASSERT.

Original-Fixed: 53654
Test: `fx test bt-host-unittest`, added "SMP_PacketTest.Mod256Equals0\
       LengthPacketGivesCorrectError" to cover the overflow case.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/400494
Original-Revision: 18ed1cb73c4070dd7586332a88bcba39705689ad
GitOrigin-RevId: 61f037b966e2381b1055293efed6c9a4ea68975c
Change-Id: Ia5d7fb67d6ca653bfc9a1a490928d33518a3f11d