[roll] Roll fuchsia [components] ElfRunner monitors for process exit
The ElfRunner waits for a ZX_PROCESS_TERMINATED signal from the
process handle for the process the runner started to host the
component. When this signal is received, it aborts the future
serving the ComponentController channel to component_manager
which causes the channel to close.

Original-Fixed: 49303

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/396072
Original-Revision: eca31c6c2ef2c86e5b5d6d66fc9db14b16dc2c06
GitOrigin-RevId: f6f6180c74948b190c06ccb900511317421e518f
Change-Id: I43fc450a8dddb6381f7cd7e27e507914a5eb2471