[roll] Roll fuchsia [audio][test] Fix WaitForPacket
After an AudioRenderer receives a Packet via SendPacket, the renderer
will hold onto the packet until it's done with that packet, at which
point the packet is returned to the caller.

It turns out that RendererShim::WaitForPacket was assuming that a packet
was completely rendered to the ring buffer by the time it is returned.
However, that may not be the case -- it may be mixed into an
intermediate buffer but not yet rendered to the final ring buffer. This
breaks tests (in a future CL) which rely on WaitForPacket to wait until
the packet is actually rendered to the ring buffer.

To fix this, I've refactored WaitForPackets such that it has enough
information to wait until the packet is fully rendered. Specifically, we
pass in the renderer's start_time (wall clock time for PTS=0) and each
packet's start_pts and size. From this info, we can compute the real
time after which the packet must have been rendered to the ring buffer.

Multiply: audio_pipeline_tests
Bugs: 49807
Original-Fixed: 53893
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/396999
Original-Revision: 358d2420b7512b7362859a0ef684f28fb6e34b99
GitOrigin-RevId: 8dff92e424d9b14449ddcfce09146bd173941c43
Change-Id: Ia1d6a293dbf43cbda829cf917e818793b30f8675