[roll] Roll fuchsia [hypervisor][arm64] Fix use of thread pointer
Just as with the x86 code, we need to make sure we are careful when we
use a thread pointer, as it could result in a use-after-free.

Update the arm64 code to track the current CPU that the thread is
executing on, just as in the x86 code.

Also make some minor tweaks to align the arm64 and x86 code.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/400254
Original-Revision: b6e2e0b59fba1be1a7ae5c78cfe84005eae5f9d7
GitOrigin-RevId: 528fa1ef342e43dede376d11b116d11428bfdc3f
Change-Id: I10ddf47ec75152ac3a2307fa85bb11713a2ccbcc