[roll] Roll fuchsia [kernel][vm] Handle ZX_ERR_INTERNAL_INTR_KILLED in page fault
If the faulting thread is killed while waiting for a page request to be
fulfilled, return successfully from the page fault handler. There's
nothing we can do in this case. This is not indicative of an error on
the page fault path.

Test: core-pager

Original-Bug: 54361
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/400582
Original-Revision: 99abae8fb155d1faede43aabac29da329c7a277b
GitOrigin-RevId: b2d035dd214afc3fad7d72b1e527fcc0cdd2037d
Change-Id: Ib6c1e2bfb737d6f75b2bfda9d567c61bef13c6b2