[roll] Roll fuchsia [debugger] Address a theoretical thread suspension race.
When a thread is requested to synchronously suspend, but there was
already an asynchronous suspend pending, wait until the suspend is
complete.

I noticed this theoretical race condition issue when looking at the
bookmark "stop mode" handling. Currently we never issue asynchronous
suspends so this issue never happens (there is one call to asynchronous
suspend, but that caller immediately waits for the thread to be
suspended). If we use suspend in more ways in the future, this might
become an issue so is worth resolving.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/400723
Original-Revision: 25c07b8729b15d612a3fb03b0958243180e7371c
GitOrigin-RevId: 6984e6ccfb73840b01ce4cd0d32f94d19fde648e
Change-Id: Iebe74218786654e33b10b6c9b7d327d6a961778f