[roll] Roll fuchsia [xhci-rewrite] Add physically contiguous TRB allocator
This patch adds a new allocator which ensures that all TRBs
returned are contiguous in physical memory.
If we are in the middle of a transfer ring and there isn't
enough contiguous space, it pads the unusable space with NOPs.

Test: fx test xhci-transfer-ring-unittest
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/400496
Original-Revision: af7b99707a9e970dfb20256fd73e23726a3216c6
GitOrigin-RevId: 50e6176d0357030d47d92825e50d403d329bab72
Change-Id: I62c806e9cdd4defc129f983a751b2f10c1951561