[roll] Roll fuchsia [cmc] Merge using serde_json rather than json5
`cmc merge` operations take >300ms on my workstation and >1sec on goma
workers. We have thousands of actions like this in our build graph so it
adds up. Though I don't believe these operations typically happen on the
critical path, it doesn't hurt to address the issue.

With this change, the same operation takes <10ms on my workstation
(`time` says it takes 0.00s of CPU and 7msec of wall time).
Don't know why serde_json is so much faster than json5 but here we are.

Technically this is a regression since now we can't merge json5.
Practically this is fine because we don't.

This seems to have a negligible effect on build time though.
$ fx set core.x64
$ fx clean-build
4145.60s user 986.40s system 2654% cpu 3:13.37
$ fx clean-build
4153.90s user 990.29s system 2685% cpu 3:11.58 total

I did multiple runs, the numbers are almost within the margin of error.

As an aside, seeing as the json5 crate does parsing fantastically slow,
and we do a lot of json5 parsing in component manager, we may want to
put more effort into this.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/401013
Original-Revision: 81b8cc541c8a5921fa5ed90aa43d5c637e9cb1ce
GitOrigin-RevId: 75b93d8c8eed9c00372d010547c8585db5c2c502
Change-Id: Ifb83b64cf0e1de7186035537123b7549c9ae5daa