[roll] Roll fuchsia [debugger] Add conditional expression support.
Adds support for C and Rust "if" statements, and C ternary ?:
expressions.

C-style if statements are not useful by themselves because our
expression system has no need to execute a sequence of steps (there is
no state). But if *expressions* (C ternary "?:" and Rust "if"
expressions) are useful for implementing both pretty-printers and, in
the future, conditional breakpoints and logging.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/398597
Original-Revision: 257296aa5040ce6ae99bb79ec061e1d16e7c22a0
GitOrigin-RevId: b6a7573e3c4f69dcf889a86ea7b6f16369ecd526
Change-Id: I562ac650297b7b2d52f267377a43e68afb1efcff