[roll] Roll fuchsia [vm][test] Fix flake in core-pager test
The CleanThreadKill test verifies that we can cleanly kill a thread
blocked on a page fault. One of the conditions it checks is that the
pager receives a page request. In some execution flows, the faulting
thread was getting killed before this check, which would also cancel the
page request, causing the test to hang waiting for a page request to
arrive.

Changed the control flow to make the sequence of events deterministic.
Got rid of a redundant thread.

Multiply: /boot/test/core-pager

Original-Bug: 54468
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/401019
Original-Revision: b7b363f533cb1f8de6afb6deacc384a33686a278
GitOrigin-RevId: c21a8f0a8d69558a7463113cd66cffac14cefcb9
Change-Id: Ifa8757169cb3f1c950b37504886febe84cd63ece