[roll] Roll fuchsia [hwstress] Run rowhammer tests for a fixed duration.
We currently run them for a fixed number of iterations, which has a wide
variation of how long it takes depending on the type of CPU the test
runs on.

Instead, just run the tests for a fixed duration.

MULTIPLY: hwstress_test
Original-Bug: 51548
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/400259
Original-Revision: 51986cf617f7527006d623da7f96d4189cadb4e5
GitOrigin-RevId: a2ec6ccfa8e4d842cb206a8d5554b3bfbd9bee81
Change-Id: Ica74be3e250eb8b030bad27a09c897f51e093291