[roll] Roll fuchsia [bt][avrcp] Fix peer disconnection logic
Currently, when a peer disconnects from AVRCP, the termination of the
AVCTP control channel task does not necessarily occur immediately. This
is because there can be outstanding requests (particularly the
hanging-get OnVolumeChanged request). Consequently, when the same peer
reconnects (i.e is not "forgotten" by the BT stack), a new AVCTP control
stream processing task is not spawned, because the original task is still
lingering.

It is expected behavior for the AVCTP control channel task to linger,
because of the outstanding hanging-get request. However, in the event
the same peer disconnects and reconnects, a new control channel
processing task should be spawned, and the old one should be cleared and
terminated.

Adds a new flag that when set, will terminate any outstanding tasks
before spawning the new task.

While we're here, update some of the log messages to be more useful.

Original-Bug: 54168
Test: Added test_peer_reconnection

Multiply: bt-avrcp-tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/399134
Original-Revision: 389b5e7cc9ea3b030c92f91b4fa221f1f09d6df3
GitOrigin-RevId: f04d15fc4bcd208ff18c69d27d377b31d710d45b
Change-Id: I9b6d52df2f48158dbe234a2cdedb93fa504e5b29