[roll] Roll fuchsia [logs] update diagnostic/logger FIDLs for user control
This change lays the foundation for the implementation of
dynamic runtime log level control, which will be implemented
as an event (specifically - a change in Interest, which
connotes a severity level) on the LogSink protocol.

The Controller protocol is removed as there are no immediate
plans for it.

Original-Bug: 54198
Test: CI/CQ to ensure no regression
      FIDL only with no implementators yet, so no specific testing

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/399537
Original-Revision: ed81e36babfded1baf95997a682c880a10364471
GitOrigin-RevId: c3a4a6648f3356ea0797fad71defcb8007da2691
Change-Id: Iaea1d7dfbd9b5c1a57e400390632d7e4634fb3a7