[roll] Roll fuchsia [msd-arm-mali] Prefer running atoms in current mode
Switching from protected to nonprotected mode takes around 500
microseconds. The code currently tries to schedule a mode switch as soon
as a it sees a runnable atom of the opposite type, so that can add up
quickly if there are very small command buffers.

Instead, prefer to execute atoms in the current mode (up to a limit of
20 in a row). This code won't reorder atoms within a job slot, but it
will select which job slot to use based on the current mode.  That's
generally enough, since if we're switching a lot then job slot 0 will be
in protected mode and job slot 1 (vertex shaders) will be in unprotected
mode.

Original-Bug: b/159041726
Original-Fixed: 13132

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/400717
Original-Revision: 7e38ba6008c9dc3fe60284d385fc71c8429cadc7
GitOrigin-RevId: 266348205ff9ba1842df9258df4249efb288d214
Change-Id: I62e13c7d160e03e5823378eeeed3105f9f3daf01