[roll] Roll fuchsia [testrunner] add ability to test over serial
This option would be exercised if FUCHSIA_SSH_KEY is not found in the
environment and FUCHSIA_SERIAL_SOCKET is: a 'serial tester' would be
initialized, which would be a thin wrapper around a client socket
connection.

This change also removes the use of ioutil.Discard (which is /dev/null)
as it was discovered that writing to that was a source of significant
slowdown.

Original-Bug: 41930

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/384014
Original-Revision: f2da102c3d9ecfb36fd7028c518e08f5b07ece95
GitOrigin-RevId: 61c80617f9f02be447ea4da9e4b34aa9b4fead3d
Change-Id: Id5992c8ad4d9e11ab251ce5c7f550eace9fd4783