[roll] Roll fuchsia [kernel][pcie] Avoid a segfault
Don't crash in the `k pciregions` kernel debug console command if the
PCIe driver has not been initialized.  Also, add a debug printf to
this command as well as the others to indicate the specific reason
that the command has done nothing in the case that the driver has not
been initialized when the command was run.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/400995
Original-Revision: 6b3b9df5f11d22e0d8c84acce51982493ed25dd1
GitOrigin-RevId: ba635388cf3d519a2cd0bbb7fffab4e25ae141ec
Change-Id: Icb18e925cda26cd53e78b42d9c008c802624b329