[roll] Roll fuchsia [vfs] Make TraversalPosition a concrete type.
The nature of TraversalPosition means that it ends up appearing all over
the codebase. Since there's currently only one kind of TraversalPosition
(AlphabeticalTraversal), replace all references to TraversalPosition
with AlphabeticalTraversal, and eliminate all the generics.

If other kinds of TraversalPosition are necessary in the future, they
should be modelled as a `struct TraversalPosition` which doesn't leak
its contained type to any classes that use it.

Original-Bug: 54231
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/399505
Original-Revision: 387032870bba49b62de5e338328fd4435449921d
GitOrigin-RevId: eecaf8cec9161fc2490af5e0204b9f8ac6bf53ea
Change-Id: I1264ee943b4d2840caafd8fc377b92ba6c6e0fe7