[roll] Roll fuchsia [bt][sm] Implement SMP Security Request
Implement the BT spec V5.0 Vol. 3 Part H 2.4.6. The request is made in
IdlePhase, but if the SMP initiator responds with a Pairing Request,
that is handled through the same flow as regular pairing in Phase 1.
Phase 1 now takes a SecurityLevel parameter instead of a MITM bool to
verify the exchanged PairingMethod. As IdlePhase now needs to handle
timeouts and be `Abort`able, it is converted to a PairingPhase.

Original-Bug: 853
Test: Run with `fx test bt-host-unittests`. Added
      SMP_ResponderPairingTest.{
        SecurityRequestCausesPairing,
        SecurityRequestWithExistingLtk,
        SecurityRequestInitiatorEncryptsWithInsufficientSecurityLtk,
        AuthenticatedSecurityRequestWithInsufficientIoCapRejected,
        HandlesMultipleSecurityRequestsCorrectly},
      SMP_IdlePhase(Death|)Test.{
        MakeInvalidSecurityRequestsDies,
        MakeEncryptedBondableSecurityRequest,
        MakeAuthenticatedNonBondableSecurityRequest} &
      SMP_Phase1Test.{
        FeatureExchangeResponderRespectsDesiredLevel,
        FeatureExchangeResponderRejectsMethodOfInsufficientSecurity}
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/392376
Original-Revision: 6cbf14304f625bc689845ba6c24e5bf8521c21ab
GitOrigin-RevId: 2f4b09b7486ccbaacc0d18888c2d65afc90d9686
Change-Id: Ic17c3360cb64db90e927357f773b87ae1153cbeb